### 2022-08-11 c9415ce
Removed SNAP_MAX & fof_com & sh_com; Search for all .hdf5 files in folder automatically
