+++ cosmic_profiles/common/profile_classes.pyx
@@ -14,7 +14,7 @@ cimport cython
 from cosmic_profiles.common.python_routines import print_status, set_axes_equal, fibonacci_ellipsoid, respectPBCNoRef, calcCoM
 from cosmic_profiles.shape_profs.shape_profs_tools import getGlobalEpsHist, getLocalEpsHist, getLocalTHist, getGlobalTHist, getShapeProfs
 from cosmic_profiles.dens_profs.dens_profs_tools import getDensProfs, fitDensProfHelper
-from cosmic_profiles.gadget_hdf5.get_hdf5 import getHDF5Data, getHDF5GxData, getHDF5SHDMData, getHDF5SHGxData, getHDF5DMData
+from cosmic_profiles.gadget_hdf5.get_hdf5 import getHDF5GxData, getHDF5SHDMData, getHDF5SHGxData, getHDF5DMData
 from cosmic_profiles.gadget_hdf5.gen_catalogues import calcCSHCat, calcGxCat
 from cosmic_profiles.cython_helpers.helper_class cimport CythonHelpers
 from cosmic_profiles.shape_profs.shape_profs_algos import calcMorphLocal, calcMorphGlobal, calcMorphLocalVelDisp, calcMorphGlobalVelDisp
@@ -1333,10 +1333,9 @@ cdef class DensProfsHDF5(CosmicBase):
     cdef str HDF5_SNAP_DEST
     cdef str HDF5_GROUP_DEST
     cdef int MIN_NUMBER_STAR_PTCS
-    cdef int SNAP_MAX
     cdef bint WANT_RVIR
     
-    def __init__(self, str HDF5_SNAP_DEST, str HDF5_GROUP_DEST, str SNAP, int SNAP_MAX, float L_BOX, int MIN_NUMBER_PTCS, int MIN_NUMBER_STAR_PTCS, str CENTER, bint WANT_RVIR):
+    def __init__(self, str HDF5_SNAP_DEST, str HDF5_GROUP_DEST, str SNAP, float L_BOX, int MIN_NUMBER_PTCS, int MIN_NUMBER_STAR_PTCS, str CENTER, bint WANT_RVIR):
         """
         :param HDF5_SNAP_DEST: where we can find the snapshot
         :type HDF5_SNAP_DEST: string
@@ -1344,8 +1343,6 @@ cdef class DensProfsHDF5(CosmicBase):
         :type HDF5_GROUP_DEST: string
         :param SNAP: e.g. '024'
         :type SNAP: string
-        :param SNAP_MAX: e.g. 16
-        :type SNAP_MAX: int
         :param SNAP: snapshot identifier, e.g. '024'
         :type SNAP: string
         :param L_BOX: simulation box side length
@@ -1364,7 +1361,6 @@ cdef class DensProfsHDF5(CosmicBase):
         self.HDF5_SNAP_DEST = HDF5_SNAP_DEST
         self.HDF5_GROUP_DEST = HDF5_GROUP_DEST
         self.MIN_NUMBER_STAR_PTCS = MIN_NUMBER_STAR_PTCS
-        self.SNAP_MAX = SNAP_MAX
         self.WANT_RVIR = WANT_RVIR
         
     def getXYZMasses(self, str obj_type = 'dm'):
@@ -1375,12 +1371,12 @@ cdef class DensProfsHDF5(CosmicBase):
         :return xyz, masses, MIN_NUMBER_PTCS: positions, masses, and minimum number of particles
         :rtype: (N2,3) floats, (N2,) floats, int"""
         if obj_type == 'dm':
-            xyz, masses, velxyz = getHDF5DMData(self.HDF5_SNAP_DEST, self.SNAP_MAX, self.SNAP)
[…]
