### 2022-08-11 10634f8
Updated 'Caching Routines' docs
