+++ cosmic_profiles/common/config.py
@@ -7,7 +7,13 @@ rank = comm.Get_rank()
 size = comm.Get_size()
 
 def initialize(GB, MAXSIZE):
-    """ Initialize all global variables, should only be called once."""
+    """ Initialize all global variables, should only be called once.
+    
+    :param GB: new memory limit in GBs
+    :type GB: int, or float
+    :param MAXSIZE: new maximum cache size (note that multiple functions are decorated, each will have
+        new maximum cache size)
+    :type MAXSIZE: int"""
     
     global GBs # Number of gigabytes. If only this much left (according to psutil.virtual_memory().available), cache will be considered full.
                # I.e. if this many GBs of memory available, stop caching to avoid thrashing.
@@ -24,7 +30,7 @@ def initialize(GB, MAXSIZE):
 def updateCachingMaxGBs(GB):
     """ Update ``use_memory_up_to`` argument of LRU-caching decorator
 
-    :param GB: new memory limit
+    :param GB: new memory limit in GBs
     :type GB: int, or float""" 
     global GBs
     GBs = GB
@@ -32,7 +38,8 @@ def updateCachingMaxGBs(GB):
 def updateCachingMaxSize(MAXSIZE):
     """ Update ``maxsize`` argument of LRU-caching decorator
 
-    :param MAXSIZE: new maximum size of each cache
+    :param MAXSIZE: new maximum cache size (note that multiple functions are decorated, each will have
+        new maximum cache size)
     :type MAXSIZE: int""" 
     global CACHE_MAXSIZE
     CACHE_MAXSIZE = MAXSIZE
