+++ docs/source/code_reference/common.rst
@@ -26,7 +26,7 @@ The ``python_routines`` module helps in calculating the center of mass of a poin
 Caching Routines
 *****************
 
-The ``caching`` module implements the caching of function outputs to avoid recalculating/reloading repeatedly. A modified decorator has been added that takes an additional optional argument ``use_memory_up_to``. If set, the cache will be considered full if there are fewer than ``use_memory_up_to bytes`` of memory available (according to ``psutil.virtual_memory().available``). Note that the ``maxsize`` argument will have no effect in that case. The motivation behind a memory-aware LRU caching function is that caching too many values causes thrashing, which should be avoided if possible.
+The ``caching`` module implements the caching of function outputs to avoid recalculating/reloading repeatedly. A modified decorator has been added that takes an additional optional argument ``use_memory_up_to``. If set (i.e. ``use_memory_up_to`` is not ``False``), the cache will be considered full if there are fewer than ``use_memory_up_to bytes`` of memory available (according to ``psutil.virtual_memory().available``). Note that the ``maxsize`` argument will have no effect in that case. The motivation behind a memory-aware LRU caching function is that caching too many values causes thrashing, which should be avoided if possible.
 
 .. automodule:: common.caching
    :members:
@@ -39,3 +39,15 @@ The user has the liberty to update ``use_memory_up_to`` and ``maxsize`` by calli
    :members:
    :undoc-members:
    :show-inheritance:
+
+The default caching parameters are
+
+.. literalinclude :: ../../../cosmic_profiles/common/config.py
+   :language: python
+   :lines: 47
+
+To disable caching, use::
+
+    from cosmic_profiles import updateCachingMaxSize, updateCachingMaxGBs
+    updateCachingMaxGBs(False)
+    updateCachingMaxSize(None)
