### 2022-08-24 4635272
For density profile estimation allow use to select halos of interest using  keyword argument; added r200 as argument for getShapeCatLocalBase, getDensProfsEllDirectBinningBase etc for overall compatibility with  keyword; updated shape_profs_algos docs, removed N vs N1 vs N2 vs N3 confusion; renamed getDensProfsDirectBinning & getDensProfsKernelBased to single estDensProfs; getDensProfs in dens_profs_tools.py renamed to plotDensProfs; float[:] r200s --> float[:] r200
