+++ cosmic_profiles/shape_profs/shape_profs_algos.pyx
@@ -523,12 +523,12 @@ def calcMorphLocal(float[:,:] xyz, float[:] masses, float[:] r200, cat, float L_
     
     :param xyz: positions of all (DM or star) particles in simulation box
     :type xyz: (N2 x 3) floats
-    :param cat: each entry of the list is a list containing indices of particles belonging to an object
-    :type cat: list of length N1
     :param masses: masses of the particles expressed in unit mass
     :type masses: (N2 x 1) floats
-    :param r200: each entry of the list gives the R_200 radius of the parent halo
-    :type r200: list of length N1
+    :param r200: R_200 radii of the parent halos
+    :type r200: (N1,) floats
+    :param cat: each entry of the list is a list containing indices of particles belonging to an object
+    :type cat: list of length N1
     :param L_BOX: simulation box side length
     :type L_BOX: float, units: Mpc/h
     :param MIN_NUMBER_PTCS: minimum number of particles for object to qualify for morphology calculation
@@ -660,12 +660,12 @@ def calcMorphGlobal(float[:,:] xyz, float[:] masses, float[:] r200, cat, float L
     
     :param xyz: positions of all (DM or star) particles in simulation box
     :type xyz: (N2 x 3) floats
-    :param cat: each entry of the list is a list containing indices of particles belonging to an object
-    :type cat: list of length N1
     :param masses: masses of the particles expressed in unit mass
     :type masses: (N2 x 1) floats
-    :param r200: each entry of the list gives the R_200 radius of the parent halo
-    :type r200: list of length N1
+    :param r200: R_200 radii of the parent halos
+    :type r200: (N1,) floats
+    :param cat: each entry of the list is a list containing indices of particles belonging to an object
+    :type cat: list of length N1
     :param L_BOX: simulation box side length
     :type L_BOX: float, units: Mpc/h
     :param MIN_NUMBER_PTCS: minimum number of particles for object to qualify for morphology calculation
@@ -790,12 +790,12 @@ def calcMorphLocalVelDisp(float[:,:] xyz, float[:,:] vxyz, float[:] masses, floa
     :type xyz: (N2 x 3) floats
     :param vxyz: velocities of all (DM or star) particles in simulation box
     :type vxyz: (N2 x 3) floats
-    :param cat: each entry of the list is a list containing indices of particles belonging to an object
-    :type cat: list of length N2
     :param masses: masses of the particles expressed in unit mass
     :type masses: (N2 x 1) floats
-    :param r200: each entry of the list gives the R_200 radius of the parent halo
-    :type r200: list of length N1
+    :param r200: R_200 radii of the parent halos
+    :type r200: (N1,) floats
+    :param cat: each entry of the list is a list containing indices of particles belonging to an object
+    :type cat: list of length N1
     :param L_BOX: simulation box side length
     :type L_BOX: float, units: Mpc/h
     :param MIN_NUMBER_PTCS: minimum number of particles for object to qualify for morphology calculation
@@ -931,12 +931,12 @@ def calcMorphGlobalVelDisp(float[:,:] xyz, float[:,:] vxyz, float[:] masses, flo
     :type xyz: (N2 x 3) floats
     :param vxyz: velocities of all (DM or star) particles in simulation box
     :type vxyz: (N2 x 3) floats
-    :param cat: each entry of the list is a list containing indices of particles belonging to an object
-    :type cat: list of length N2
     :param masses: masses of the particles expressed in unit mass
     :type masses: (N2 x 1) floats
     :param r200: R_200 radii of the parent halos
[…]
