+++ cosmic_profiles/tests/test_densities_ex_script.py
@@ -63,8 +63,9 @@ def test_densities_ex_script():
     
     ############################## Estimate Density Profile ####################################################
     # Visualize density profile: A sample output is shown above!
-    dens_profs_db = cprofiles.getDensProfsDirectBinning(r_over_rvir) # dens_profs_db is in M_sun*h^2/Mpc^3
-    dens_profs_kb = cprofiles.getDensProfsKernelBased(r_over_rvir)
+    halos_select = [0, 0]
+    dens_profs_db = cprofiles.estDensProfs(r_over_rvir, select = halos_select, direct_binning = True) # dens_profs_db is in M_sun*h^2/Mpc^3
+    dens_profs_kb = cprofiles.estDensProfs(r_over_rvir, select = halos_select, direct_binning = False)
     plt.figure()
     plt.loglog(r_over_rvir, dens_profs_db[0], 'o--', label='direct binning', markersize = 3)
     plt.loglog(r_over_rvir, dens_profs_kb[0], 'o--', label='kernel-based', markersize = 3)
@@ -77,7 +78,7 @@ def test_densities_ex_script():
     ############################## Fit Density Profile #########################################################
     r_over_rvir_fit = r_over_rvir[10:] # Do not fit innermost region since not reliable in practice. Use gravitational softening scale and / or relaxation timescale to estimate inner convergence radius.
     dens_profs_db_fit = dens_profs_db[0,10:]
-    best_fit = cprofiles.getDensProfsBestFits(dens_profs_db_fit.reshape((1,dens_profs_db_fit.shape[0])), r_over_rvir_fit, method = 'einasto')
+    best_fit = cprofiles.fitDensProfs(dens_profs_db_fit.reshape((1,dens_profs_db_fit.shape[0])), r_over_rvir_fit, method = 'einasto')
     plt.figure()
     plt.loglog(r_over_rvir, dens_profs_db[0], 'o--', label='density profile', markersize = 4)
     model_pars = {'rho_s': best_fit[0,0], 'alpha': best_fit[0,1], 'r_s': best_fit[0,2]}
