### 2022-08-25 3c2db06
Raise ValueError if  argument is incompatible with  in estConcentrations() etc; updated for_docs and docs folders
