+++ cosmic_profiles/for_docs/common/profile_classes.pyx
@@ -18,13 +18,12 @@ rank = comm.Get_rank()
 size = comm.Get_size()
 
 @cython.embedsignature(True)
-@cython.binding(True)
 cdef class CosmicBase:
     """ Parent class governing high-level cosmic shape calculations
     
     Its public methods are ``getR200s()``, ``getMassesCentersBase()``, 
     ``getShapeCatLocalBase()``, ``getShapeCatGlobalBase()``, ``getShapeCatVelLocalBase()``, 
-    ``getShapeCatVelGlobalBase()``, ``dumpShapeCatLocalBase()``, ``dumpShapeCatGlobalBase()``, 
+    ``getShapeCatVelGlobalBase()``, ``dumpShapeCatLocalBase()``, ``dumpShapeCatGlobalBase()``,
     ``dumpShapeCatVelLocalBase()``, ``dumpShapeCatVelGlobalBase()``, ``plotShapeProfsBase()``,
     ``plotLocalTHistBase()``, ``plotGlobalTHistBase()``, ``getDensProfsBestFitsBase()``,
     ``getConcentrationsBase()``, ``getDensProfsSphDirectBinningBase()``, ``getDensProfsEllDirectBinningBase()``,
@@ -77,13 +76,15 @@ cdef class CosmicBase:
         :rtype: (N,3) and (N,) floats"""
         return
     
-    def getShapeCatLocalBase(self, float[:,:] xyz, float[:] masses, idx_cat, int MIN_NUMBER_PTCS, float D_LOGSTART, float D_LOGEND, int D_BINS, float M_TOL, int N_WALL, int N_MIN, bint reduced, bint shell_based):
+    def getShapeCatLocalBase(self, float[:,:] xyz, float[:] masses, float[:] r200, idx_cat, int MIN_NUMBER_PTCS, float D_LOGSTART, float D_LOGEND, int D_BINS, float M_TOL, int N_WALL, int N_MIN, bint reduced, bint shell_based):
         """ Get all relevant local shape data
         
         :param xyz: positions of all simulation particles
         :type xyz: (N2,3) floats, N2 >> N1
         :param masses: masses of all simulation particles
         :type masses: (N2,) floats
+        :param r200: each entry gives the R_200 radius of the parent halo
+        :type r200: (N1,) floats
         :param idx_cat: each entry of the list is a list containing indices of particles belonging to an object
         :type idx_cat: list of length N1
         :param MIN_NUMBER_PTCS: minimum number of particles for object to qualify for morphology calculation
@@ -113,13 +114,15 @@ cdef class CosmicBase:
         """
         return
     
-    def getShapeCatGlobalBase(self, float[:,:] xyz, float[:] masses, idx_cat, int MIN_NUMBER_PTCS, float M_TOL, int N_WALL, int N_MIN, bint reduced):
+    def getShapeCatGlobalBase(self, float[:,:] xyz, float[:] masses, float[:] r200, idx_cat, int MIN_NUMBER_PTCS, float M_TOL, int N_WALL, int N_MIN, bint reduced):
         """ Get all relevant global shape data
         
         :param xyz: positions of all simulation particles
         :type xyz: (N2,3) floats, N2 >> N1
         :param masses: masses of all simulation particles
         :type masses: (N2,) floats
+        :param r200: each entry gives the R_200 radius of the parent halo
+        :type r200: (N1,) floats
         :param idx_cat: each entry of the list is a list containing indices of particles belonging to an object
         :type idx_cat: list of length N1
         :param MIN_NUMBER_PTCS: minimum number of particles for object to qualify for morphology calculation
@@ -141,7 +144,7 @@ cdef class CosmicBase:
         """
         return
         
[…]
