### 2022-08-25 10e2004
Fixed getHDF5DMData() etc bug in arguments; Fixed hdf5SnapFilenamesList etc bug; Improved construction of h_cat and gx_cat list; Removed in degrees comment in print statement; Fixed plotDensProfs() bug, getIdxCat must come before len(idx_cat); M_split print statements removed; Allowing user to choose number of mass bins; Renamed N_MIN etc to IT_MIN; Updated gadget_hdf5.py etc
