+++ cosmic_profiles/gadget_hdf5/get_hdf5.py
@@ -39,7 +39,7 @@ def getHDF5Data(HDF5_SNAP_DEST, HDF5_GROUP_DEST):
         hdf5SnapFilenamesList = glob.glob('{}/*.hdf5'.format(HDF5_SNAP_DEST))
         hdf5GroupFilenamesList = glob.glob('{}/*.hdf5'.format(HDF5_GROUP_DEST))
         for fname in hdf5GroupFilenamesList:
-            g = h5py.File(r'{}/{}'.format(HDF5_GROUP_DEST, fname), 'r')
+            g = h5py.File(r'{}'.format(fname), 'r')
             if 'Group/GroupLenType' in g:
                 to_be_appended = [np.float32(g['Group/GroupLenType'][i,1]) for i in range(g['Group/GroupLenType'].shape[0])]
                 fof_dm_sizes.append(to_be_appended)
@@ -48,7 +48,7 @@ def getHDF5Data(HDF5_SNAP_DEST, HDF5_GROUP_DEST):
             if 'Subhalo/SubhaloLenType' in g:
                 sh_len.append([np.float32(g['Subhalo/SubhaloLenType'][i,1]) for i in range(g['Subhalo/SubhaloLenType'].shape[0])])
         for fname in hdf5SnapFilenamesList:
-            f = h5py.File(r'{}/{}'.format(HDF5_SNAP_DEST, fname), 'r')
+            f = h5py.File(r'{}'.format(fname), 'r')
             dm_x = np.hstack((dm_x, np.float32(f['PartType1/Coordinates'][:,0]/1000))) # in Mpc = 3.085678e+27 cm
             dm_y = np.hstack((dm_y, np.float32(f['PartType1/Coordinates'][:,1]/1000))) 
             dm_z = np.hstack((dm_z, np.float32(f['PartType1/Coordinates'][:,2]/1000))) 
@@ -83,7 +83,9 @@ def getHDF5GxData(HDF5_SNAP_DEST, HDF5_GROUP_DEST):
     :type HDF5_GROUP_DEST: string
     :return: star_xyz (star ptc positions),
         nb_shs (# subhalos in each FoF-halo), 
-        star_masses (mass of each star ptc)
+        star_masses (mass of each star ptc),
+        star_velxyz (star ptc velocities),
+        is_star (whether or not particle is star particle or wind particle)
     :rtype: float and int arrays"""
     def inner(HDF5_SNAP_DEST, HDF5_GROUP_DEST):
         is_star = np.empty(0, dtype = bool)
@@ -107,7 +109,7 @@ def getHDF5GxData(HDF5_SNAP_DEST, HDF5_GROUP_DEST):
         else:
             last = rank == nb_jobs_to_do - 1
         for snap_run in range(rank*perrank, rank*perrank+do_sth*(perrank+last*(nb_jobs_to_do-(rank+1)*perrank))):
-            f = h5py.File(r'{}/{}'.format(HDF5_SNAP_DEST, hdf5SnapFilenamesList[snap_run]), 'r')
+            f = h5py.File(r'{}'.format(hdf5SnapFilenamesList[snap_run]), 'r')
             if 'PartType4/Coordinates' in f:
                 is_star = np.hstack((is_star, f['PartType4/GFM_StellarFormationTime'][:] > 0)) # Wind particle?
                 star_x = np.hstack((star_x, np.float32(f['PartType4/Coordinates'][:,0]/1000))) # in Mpc = 3.085678e+27 cm
@@ -197,7 +199,7 @@ def getHDF5GxData(HDF5_SNAP_DEST, HDF5_GROUP_DEST):
         else:
             last = rank == nb_jobs_to_do - 1
[…]
