+++ cosmic_profiles/shape_profs/shape_profs_algos.pyx
@@ -13,7 +13,7 @@ cimport cython
 from libc.math cimport sqrt
 
 @cython.embedsignature(True)
-cdef float[:] runShellAlgo(float[:] morph_info, float[:,:] xyz, float[:,:] xyz_princ, float[:] masses, int[:] shell, float[:] r_ell, float[:] center, complex[::1,:] shape_tensor, double[::1] eigval, complex[::1,:] eigvec, float d, float delta_d, float M_TOL, int N_WALL, int N_MIN, bint reduced) nogil:
+cdef float[:] runShellAlgo(float[:] morph_info, float[:,:] xyz, float[:,:] xyz_princ, float[:] masses, int[:] shell, float[:] r_ell, float[:] center, complex[::1,:] shape_tensor, double[::1] eigval, complex[::1,:] eigvec, float d, float delta_d, float IT_TOL, int IT_WALL, int IT_MIN, bint reduced) nogil:
     """ S1 algorithm for halos/galaxies at elliptical radius ``d`` with shell width ``delta_d``
     
     Calculates the axis ratios at a distance ``d`` from the center of the entire particle distro.\n
@@ -51,14 +51,14 @@ cdef float[:] runShellAlgo(float[:] morph_info, float[:,:] xyz, float[:,:] xyz_p
     :type d: float
     :param delta_d: thickness of the shell in real space (constant across shells in logarithmic space)
     :type delta_d: float
-    :param M_TOL: convergence tolerance, eigenvalue fractions must differ by less than ``M_TOL``
+    :param IT_TOL: convergence tolerance, eigenvalue fractions must differ by less than ``IT_TOL``
         for iteration to stop
-    :type M_TOL: float
-    :param N_WALL: maximum permissible number of iterations
-    :type N_WALL: float
-    :param N_MIN: minimum number of particles (DM or star particle) in any iteration; 
+    :type IT_TOL: float
+    :param IT_WALL: maximum permissible number of iterations
+    :type IT_WALL: float
+    :param IT_MIN: minimum number of particles (DM or star particle) in any iteration; 
         if undercut, shape is unclassified
-    :type N_MIN: int
+    :type IT_MIN: int
     :param reduced: whether or not reduced shape tensor (1/r^2 factor)
     :type reduced: boolean
     :return: ``morph_info`` containing d, q, s, eigframe info
@@ -85,11 +85,11 @@ cdef float[:] runShellAlgo(float[:] morph_info, float[:,:] xyz, float[:,:] xyz_p
         else:
             corr += 1
         r_ell[i] = sqrt((center[0]-xyz[i,0])**2+(center[1]-xyz[i,1])**2/q_new**2+(center[2]-xyz[i,2])**2/s_new**2)
-    while (err > M_TOL):
-        if iteration > N_WALL:
+    while (err > IT_TOL):
+        if iteration > IT_WALL:
             morph_info[:] = 0.0
             return morph_info
-        if pts_in_shell < N_MIN:
+        if pts_in_shell < IT_MIN:
             morph_info[:] = 0.0
             return morph_info
         # Get shape tensor
@@ -145,7 +145,7 @@ cdef float[:] runShellAlgo(float[:] morph_info, float[:,:] xyz, float[:,:] xyz_p
     return morph_info
 
 @cython.embedsignature(True)
[…]
