### 2022-08-25 f5c52be
Updated shape estimation docs
