### 2022-08-25 686c3cd
Fixed mass bins for drawDensProfs()
