+++ cosmic_profiles/dens_profs/dens_profs_tools.py
@@ -75,7 +75,7 @@ def getHernquistProf(r, model_pars):
     r_s = model_pars['r_s']
     return rho_s/((r/r_s)*(1+r/r_s)**3)
 
-def drawDensProfs(VIZ_DEST, SNAP, cat, r200s, dens_profs_fit, ROverR200_fit, dens_profs, ROverR200, obj_masses, obj_centers, method, nb_bins, start_time, MASS_UNIT, suffix = '_'):
+def drawDensProfs(VIZ_DEST, SNAP, cat, r200s, dens_profs_fit, ROverR200_fit, dens_profs, ROverR200, obj_masses, obj_centers, method, nb_bins, select, start_time, MASS_UNIT, suffix = '_'):
     """
     Create a series of plots to analyze object shapes
     
@@ -106,6 +106,8 @@ def drawDensProfs(VIZ_DEST, SNAP, cat, r200s, dens_profs_fit, ROverR200_fit, den
     :type method: string, either `einasto`, `alpha_beta_gamma`, `hernquist`, `nfw`
     :param nb_bins: Number of mass bins to plot density profiles for
     :type nb_bins: int
+    :param select: index of first and last object to look at in the format [idx_first, idx_last]
+    :type select: list containing two integers
     :param start_time: time of start of shape analysis
     :type start_time: float
     :param MASS_UNIT: conversion factor from previous mass unit to M_sun/h
@@ -128,14 +130,14 @@ def drawDensProfs(VIZ_DEST, SNAP, cat, r200s, dens_profs_fit, ROverR200_fit, den
             return model_pars_dict
         
         # Mass splitting
-        max_min_m, obj_m_groups, obj_center_groups, idx_groups = M_split(MASS_UNIT*obj_masses, obj_centers, start_time, NB_BINS = nb_bins)
+        max_min_m, obj_m_groups, obj_center_groups, idx_groups = M_split(MASS_UNIT*obj_masses[select[0]:select[1]+1], obj_centers[select[0]:select[1]+1], start_time, NB_BINS = nb_bins)
         del obj_centers; del obj_center_groups; del idx_groups
         obj_pass = np.int32(np.array([1 if x != [] else 0 for x in cat]))
-        print_status(rank, start_time, "The number of objects considered is {0}".format(np.sum(obj_pass)))
-        print_status(rank, start_time, "The mass bins (except maybe last) have size {0}".format(len(obj_m_groups[0])))
-        print_status(rank, start_time, "The number of mass bins is {0}".format(len(obj_m_groups)))
         idxs_compr = np.zeros((len(cat),), dtype = np.int32)
         idxs_compr[obj_pass.nonzero()[0]] = np.arange(np.sum(obj_pass)) 
+        print_status(rank, start_time, "The number of objects considered is {0}".format(np.sum(obj_pass)))
+        print_status(rank, start_time, "The mass bins (except maybe last) have size {0}".format(np.sum(np.int32([1 if (obj_pass[i] == 1 and obj_masses[idxs_compr[i]]*MASS_UNIT > max_min_m[1] and obj_masses[idxs_compr[i]]*MASS_UNIT < max_min_m[2]) else 0 for i in range(len(cat))]))))
+        print_status(rank, start_time, "The number of mass bins is {0}".format(len(obj_m_groups)))
         prof_models = {'einasto': getEinastoProf, 'alpha_beta_gamma': getAlphaBetaGammaProf, 'nfw': getNFWProf, 'hernquist': getHernquistProf}
         model_name = {'einasto': 'Einasto', 'alpha_beta_gamma': r'\alpha \beta \gamma', 'nfw': 'NFW', 'hernquist': 'Hernquist'}
         
