+++ cosmic_profiles/for_docs/common/cosmic_base_class.pyx
@@ -1,0 +1,595 @@
+#cython: language_level=3
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+import time
+cimport cython
+from mpi4py import MPI
+comm = MPI.COMM_WORLD
+rank = comm.Get_rank()
+size = comm.Get_size()
+
+@cython.embedsignature(True)
+cdef class CosmicBase:
+    """ Parent class governing high-level cosmic shape calculations
+    
+    Its public methods are ``getR200s()``, ``getMassesCentersBase()``, 
+    ``getShapeCatLocalBase()``, ``getShapeCatGlobalBase()``, ``getShapeCatVelLocalBase()``, 
+    ``getShapeCatVelGlobalBase()``, ``dumpShapeCatLocalBase()``, ``dumpShapeCatGlobalBase()``,
+    ``dumpShapeCatVelLocalBase()``, ``dumpShapeCatVelGlobalBase()``, ``plotShapeProfsBase()``,
+    ``plotLocalTHistBase()``, ``plotGlobalTHistBase()``, ``getDensProfsBestFitsBase()``,
+    ``getConcentrationsBase()``, ``getDensProfsSphDirectBinningBase()``, ``getDensProfsEllDirectBinningBase()``,
+    ``getDensProfsKernelBasedBase()``, ``getObjInfoBase()``"""
+    
+    def __init__(self, str SNAP, float L_BOX, int MIN_NUMBER_PTCS, str CENTER):
+        """
+        :param SNAP: snapshot identifier, e.g. '024'
+        :type SNAP: string
+        :param L_BOX: simulation box side length
+        :type L_BOX: float, units: Mpc/h
+        :param MIN_NUMBER_PTCS: minimum number of particles for object to qualify for morphology calculation
+        :type MIN_NUMBER_PTCS: int
+        :param CENTER: shape quantities will be calculated with respect to CENTER = 'mode' (point of highest density)
+            or 'com' (center of mass) of each halo
+        :type CENTER: str"""
+        self.SNAP = SNAP
+        self.L_BOX = L_BOX
+        self.CENTER = CENTER
+        self.MIN_NUMBER_PTCS = MIN_NUMBER_PTCS
+        self.start_time = time.time()
+        self.SAFE = 6
+        self.MASS_UNIT = 1e10
+        self.r200 = None
+    
+    def getR200s(self):
+        """ Get overdensity radii"""
+        return
+        
+    def getMassesCentersBase(self, float[:,:] xyz, float[:] masses, idx_cat, int MIN_NUMBER_PTCS):
+        """ Calculate total mass and centers of objects
+        
+        :param xyz: positions of all simulation particles
+        :type xyz: (N2,3) floats, N2 >> N1
+        :param masses: masses of all simulation particles
+        :type masses: (N2,) floats
+        :param idx_cat: each entry of the list is a list containing indices of particles belonging to an object
+        :type idx_cat: list of length N1
+        :param MIN_NUMBER_PTCS: minimum number of particles for object to qualify for morphology calculation
+        :type MIN_NUMBER_PTCS: int
+        :return centers, m: centers and masses
+        :rtype: (N,3) and (N,) floats"""
+        return
+    
+    def getShapeCatLocalBase(self, float[:,:] xyz, float[:] masses, float[:] r200, idx_cat, int MIN_NUMBER_PTCS, float D_LOGSTART, float D_LOGEND, int D_BINS, float IT_TOL, int IT_WALL, int IT_MIN, bint reduced, bint shell_based):
+        """ Get all relevant local shape data
+        
+        :param xyz: positions of all simulation particles
+        :type xyz: (N2,3) floats, N2 >> N1
+        :param masses: masses of all simulation particles
+        :type masses: (N2,) floats
[…]
