+++ docs/cythonize.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 
-cd ../cosmic_profiles
-python3 setup.py build_ext --inplace
-cd ../docs
+cd ../
+python3 setup_compile.py build_ext --inplace
+cd docs
