+++ docs/run_doc.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 cd ../cosmic_profiles/for_docs
-python3 setup.py build_ext --inplace
+python3 setup_compile.py build_ext --inplace
 cd ../../docs
 sphinx-build -b html source build
 cd ../cosmic_profiles/for_docs/common
