+++ .github/workflows/tests.yml
@@ -38,9 +38,9 @@ jobs:
           # The GitHub editor is 127 chars wide and exit-zero treats all errors as warnings.
           $CONDA/bin/flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
       - name: Compile Cython modules
-        working-directory: ./cosmic_profiles
+        working-directory: ./
         run: |
-          $CONDA/bin/python setup.py build_ext --inplace
+          $CONDA/bin/python setup_compile.py build_ext --inplace
       - name: Test with pytest
         working-directory: ./
         run: |
