### 2022-08-26 da9256f
Removed redundant idx_cat_list to idx_cat_arr conversions
