+++ cosmic_profiles/common/cosmic_base_class.pyx
@@ -56,23 +56,23 @@ cdef class CosmicBase:
         else:
             return None
         
-    def getMassesCentersBase(self, float[:,:] xyz, float[:] masses, idx_cat, int MIN_NUMBER_PTCS):
+    def getMassesCentersBase(self, float[:,:] xyz, float[:] masses, int[:,:] idx_cat, int[:] obj_size):
         """ Calculate total mass and centers of objects
         
         :param xyz: positions of all simulation particles
         :type xyz: (N2,3) floats, N2 >> N1
         :param masses: masses of all simulation particles
         :type masses: (N2,) floats
-        :param idx_cat: each entry of the list is a list containing indices of particles belonging to an object
-        :type idx_cat: list of length N1
-        :param MIN_NUMBER_PTCS: minimum number of particles for object to qualify for morphology calculation
-        :type MIN_NUMBER_PTCS: int
+        :param idx_cat: each row contains indices of particles belonging to an object
+        :type idx_cat: (N1, N3) integers
+        :param obj_size: indicates how many particles are in each object
+        :type obj_size: (N1,) integers
         :return centers, m: centers and masses
         :rtype: (N,3) and (N,) floats"""
-        centers, m = calcMassesCenters(xyz.base, masses.base, idx_cat, MIN_NUMBER_PTCS, self.L_BOX, self.CENTER)
+        centers, m = calcMassesCenters(xyz.base, masses.base, idx_cat, obj_size, self.L_BOX, self.CENTER)
         return centers, m
     
-    def getShapeCatLocalBase(self, float[:,:] xyz, float[:] masses, float[:] r200, idx_cat, int MIN_NUMBER_PTCS, float D_LOGSTART, float D_LOGEND, int D_BINS, float IT_TOL, int IT_WALL, int IT_MIN, bint reduced, bint shell_based):
+    def getShapeCatLocalBase(self, float[:,:] xyz, float[:] masses, float[:] r200, int[:,:] idx_cat, int[:] obj_size, float D_LOGSTART, float D_LOGEND, int D_BINS, float IT_TOL, int IT_WALL, int IT_MIN, bint reduced, bint shell_based):
         """ Get all relevant local shape data
         
         :param xyz: positions of all simulation particles
@@ -81,10 +81,10 @@ cdef class CosmicBase:
         :type masses: (N2,) floats
         :param r200: each entry gives the R_200 radius of the parent halo
         :type r200: (N1,) floats
-        :param idx_cat: each entry of the list is a list containing indices of particles belonging to an object
-        :type idx_cat: list of length N1
-        :param MIN_NUMBER_PTCS: minimum number of particles for object to qualify for morphology calculation
-        :type MIN_NUMBER_PTCS: int
+        :param idx_cat: each row contains indices of particles belonging to an object
+        :type idx_cat: (N1, N3) integers
+        :param obj_size: indicates how many particles are in each object
+        :type obj_size: (N1,) integers
         :param D_LOGSTART: logarithm of minimum ellipsoidal radius of interest, in units of R200 of parent halo
         :type D_LOGSTART: int
         :param D_LOGEND: logarithm of maximum ellipsoidal radius of interest, in units of R200 of parent halo
@@ -109,12 +109,12 @@ cdef class CosmicBase:
             (number_of_objs,3) float array, (number_of_objs,) float array
         """
         if rank == 0:
[…]
