### 2022-08-26 6ff749d
Got rid of -1 shift in index catalogue calculation which was needed for idx_cat_arr to idx_cat_list conversion previously; Updated tests*.py to account for modified setup_compile.py
