### 2022-08-26 3aba1c4
Fixed mpi process handling in test scripts; removed getIdxCorr() function
