+++ cosmic_profiles/tests/test_densities_ex_script.py
@@ -21,6 +21,10 @@ from cosmic_profiles import genHalo, DensProfs, getEinastoProf
 import matplotlib.pyplot as plt
 import matplotlib
 matplotlib.rcParams.update({'font.size': 13})
+from mpi4py import MPI
+comm = MPI.COMM_WORLD
+rank = comm.Get_rank()
+size = comm.Get_size()
 
 def test_densities_ex_script():
     
@@ -66,24 +70,35 @@ def test_densities_ex_script():
     halos_select = [0, 0]
     dens_profs_db = cprofiles.estDensProfs(r_over_rvir, select = halos_select, direct_binning = True) # dens_profs_db is in M_sun*h^2/Mpc^3
     dens_profs_kb = cprofiles.estDensProfs(r_over_rvir, select = halos_select, direct_binning = False)
-    plt.figure()
-    plt.loglog(r_over_rvir, dens_profs_db[0], 'o--', label='direct binning', markersize = 3)
-    plt.loglog(r_over_rvir, dens_profs_kb[0], 'o--', label='kernel-based', markersize = 3)
-    plt.loglog(r_over_rvir, getEinastoProf(r_over_rvir*r_vir[0], model_pars), lw = 1.0, label=r'Einasto-target: $\alpha$ = {:.2f}, $r_s$ = {:.2f} cMpc/h'.format(alpha, r_s))
-    plt.xlabel(r'r/$R_{\mathrm{vir}}$')
-    plt.ylabel(r"$\rho$ [$h^2M_{{\odot}}$ / Mpc${{}}^3$]")
-    plt.legend(fontsize="small", loc='lower left')
-    plt.savefig('{}/RhoProfObj0_{}.pdf'.format(VIZ_DEST, SNAP), bbox_inches='tight')
+    
+    if rank == 0:
+        dens_profs_db = dens_profs_db[0]
+        dens_profs_kb = dens_profs_kb[0]
+        plt.figure()
+        plt.loglog(r_over_rvir, dens_profs_db[0], 'o--', label='direct binning', markersize = 3)
+        plt.loglog(r_over_rvir, dens_profs_kb[0], 'o--', label='kernel-based', markersize = 3)
+        plt.loglog(r_over_rvir, getEinastoProf(r_over_rvir*r_vir[0], model_pars), lw = 1.0, label=r'Einasto-target: $\alpha$ = {:.2f}, $r_s$ = {:.2f} cMpc/h'.format(alpha, r_s))
+        plt.xlabel(r'r/$R_{\mathrm{vir}}$')
+        plt.ylabel(r"$\rho$ [$h^2M_{{\odot}}$ / Mpc${{}}^3$]")
+        plt.legend(fontsize="small", loc='lower left')
+        plt.savefig('{}/RhoProfObj0_{}.pdf'.format(VIZ_DEST, SNAP), bbox_inches='tight')
+    else:
+        dens_profs_db = np.zeros((r_over_rvir.shape[0]), dtype = np.float32)
+        dens_profs_db = np.zeros((r_over_rvir.shape[0]), dtype = np.float32)
+    comm.Bcast(dens_profs_db, root = 0)
+    comm.Bcast(dens_profs_kb, root = 0)
     
     ############################## Fit Density Profile #########################################################
     r_over_rvir_fit = r_over_rvir[10:] # Do not fit innermost region since not reliable in practice. Use gravitational softening scale and / or relaxation timescale to estimate inner convergence radius.
-    dens_profs_db_fit = dens_profs_db[0,10:]
[…]
