### 2022-08-26 b2a4e9d
Filled missing occurrences of .base; updated example_scripts
