### 2022-08-26 5eefc55
Fixed bug in dens_profs_classes.pyx, MIN_NUMBER_PTCS was redundant argument handed over to getMassesCentersBase; fixed bug in shape_profs_classes.pyx, self.idx_cat need self.idx_cat[select..], obj_size missing when handing over to getLocalShapeCatBase()
