### 2022-08-26 84c0c60
Fixed getShapeCatLocalBase() call in vizLocalShapes()
