### 2022-08-26 a2a3def
Added .base in calcMassesCenters() function call in getMassesCentersBase()
