### 2022-08-26 #4626 @adonath
Thanks @tibordome for the submission! I recently joined the editor team of JOSS and I'm looking forward to take the editing role here.
