### 2022-08-26 9f4ebdc
Fixed indexing error in vizLocalShapes() and vizGlobalShapes()
