### 2022-08-26 72a0f99
Fixed nb_suff_res in test_shapes.py; parallelized getHDF5SHDMData() and getHDF5GxDMData()
