+++ cosmic_profiles/gadget_hdf5/get_hdf5.py
@@ -228,35 +228,97 @@ def getHDF5SHDMData(HDF5_GROUP_DEST, WANT_RVIR):
             with respect to the virial radius R_vir or the overdensity radius R_200
     :type WANT_RVIR: boolean
     :return: nb_shs (# subhalos in each FoF-halo), sh_len (size of each SH), 
-        fof_dm_sizes (size of each FoF-halo), group_r200 (R200 radius of each FoF-halo), 
+        fof_dm_sizes (size of each FoF-halo), group_r200 (R200 radius of each FoF-halo),
         fof_masses (mass of each FoF-halo)
     :rtype: float and int arrays"""
     def inner(HDF5_GROUP_DEST, WANT_RVIR):
-        fof_dm_sizes = []
-        nb_shs = []
-        sh_len = []
+        fof_dm_sizes = np.empty(0, dtype = np.int32)
+        nb_shs = np.empty(0, dtype = np.int32)
+        sh_len = np.empty(0, dtype = np.int32)
         group_r200 = np.empty(0, dtype = np.float32)
         fof_masses = np.empty(0, dtype = np.float32)
         hdf5GroupFilenamesList = glob.glob('{}/*.hdf5'.format(HDF5_GROUP_DEST))
-        for fname in hdf5GroupFilenamesList:
-            g = h5py.File(r'{}'.format(fname), 'r')
+        nb_jobs_to_do = len(hdf5GroupFilenamesList)
+        perrank = nb_jobs_to_do//size + (nb_jobs_to_do//size == 0)*1
+        do_sth = rank <= nb_jobs_to_do-1
+        if size <= nb_jobs_to_do:
+            last = rank == size - 1 # Whether or not last process
+        else:
+            last = rank == nb_jobs_to_do - 1
+        count_fof = 0
+        count_sh = 0
+        for snap_run in range(rank*perrank, rank*perrank+do_sth*(perrank+last*(nb_jobs_to_do-(rank+1)*perrank))):
+            g = h5py.File(r'{}'.format(hdf5GroupFilenamesList[snap_run]), 'r')
             if 'Group/GroupLenType' in g:
-                to_be_appended = [np.int32(g['Group/GroupLenType'][i,1]) for i in range(g['Group/GroupLenType'].shape[0])]
+                fof_dm_sizes = np.hstack((fof_dm_sizes, np.int32([np.int32(g['Group/GroupLenType'][i,1]) for i in range(g['Group/GroupLenType'].shape[0])])))
                 if WANT_RVIR:
                     group_r200 = np.hstack((group_r200, np.float32(g['Group/Group_R_TopHat200'][:]/1000)))
                 else:
                     group_r200 = np.hstack((group_r200, np.float32(g['Group/Group_R_Mean200'][:]/1000)))
                 fof_masses = np.hstack((fof_masses, np.float32(g['Group/GroupMassType'][:,1])))
-                nb_shs.append([np.int32(g['Group/GroupNsubs'][i]) for i in range(g['Group/GroupNsubs'].shape[0])])
-            else:
-                to_be_appended = []
+                nb_shs = np.hstack((nb_shs, np.int32([np.int32(g['Group/GroupNsubs'][i]) for i in range(g['Group/GroupNsubs'].shape[0])])))
+                count_fof += g['Group/GroupLenType'].shape[0]
             if 'Subhalo/SubhaloLenType' in g:
[…]
