### 2022-08-26 fe4c5e9
Updated docs to properly reflect meaning of select argument
