+++ docs/source/density_profiles/index.rst
@@ -18,9 +18,9 @@ where :math:`(x,y,z)` are the coordinates of a point cloud particle in some coor
 
 To estimate density profiles with *Cosmic Profiles*, we first instantiate a ``DensProfs`` object called ``cprofiles``, similar to what we saw in the :ref:`Shape Estimation section<Shape Estimation>`. Now we can simply invoke the command::
 
-    dens_profs_db = cprofiles.estDensProfs(r_over_r200, select = [0, nb_objs-1], direct_binning = True, spherical = True),
+    dens_profs_db = cprofiles.estDensProfs(r_over_r200, select = [0, 9], direct_binning = True, spherical = True),
 
-where the float array ``dens_profs_db`` of shape :math:`(N_{\text{pass}}, N_r)` contains the estimated density profiles. The ``select`` argument expects a list of two integers indicating for which objects to estimate the density profile. In the example above, all objects are considered since the range ``[0, nb_objs-1]`` spans the maximum range of indices. As in the :ref:`Shape Estimation section<Shape Estimation>`, :math:`N_{\text{pass}}` stands for the number of objects that have been selected with the ``select`` argument and in addition are sufficiently resolved. This assumes that the float array that specifies for which unitless spherical radii ``r_over_r200`` the local density should be calculated has shape :math:`N_r`. Specifying radial bins with equal spacing in logarithmic space :math:`\log (\delta r/r_{200}) = \mathrm{const}` is common practice.
+where the float array ``dens_profs_db`` of shape :math:`(N_{\text{pass}}, N_r)` contains the estimated density profiles. The ``select`` argument expects a list of two integers indicating for which objects to estimate the density profile. In the example above, only the first 10 objects that have sufficient resolution will be considered. As in the :ref:`Shape Estimation section<Shape Estimation>`, :math:`N_{\text{pass}}` stands for the number of objects that have been selected with the ``select`` argument and in addition are sufficiently resolved. This assumes that the float array that specifies for which unitless spherical radii ``r_over_r200`` the local density should be calculated has shape :math:`N_r`. Specifying radial bins with equal spacing in logarithmic space :math:`\log (\delta r/r_{200}) = \mathrm{const}` is common practice.
 
 .. note:: In case of a Gadget-style HDF5 snapshot output, you may want to additionally specify ``obj_type = 'dm'`` or ``'gx'`` to ``estDensProfs()`` in order to have the density profiles calculated for either dark matter halos or galaxies.
 
@@ -59,7 +59,7 @@ Finally, the :math:`\alpha \beta \gamma` density profile (`Zemp et al 2011 <http
 
 To fit density profiles according to model ``method``, a string which can be either ``nfw``, ``hernquist``, ``einasto`` or ``alpha_beta_gamma``, invoke the method::
 
-    best_fits = cprofiles.fitDensProfs(dens_profs_fit, r_over_r200_fit, method, select = [0, nb_objs-1]).
+    best_fits = cprofiles.fitDensProfs(dens_profs_fit, r_over_r200_fit, method, select = [0, 9]).
 
[…]
