### 2022-08-26 ef64eee
Fixed self.MIN_NUMBER_PTCS occurrence in self.getShapeCatGlobalBase
