### 2022-08-27 639d02c
Moved getIdxCat(obj_type) out of rank == 0 environments; same for getHDF5SHDMData() etc
