+++ cosmic_profiles/dens_profs/dens_profs_classes.pyx
@@ -278,32 +278,38 @@ cdef class DensProfsHDF5(CosmicBase):
             obj_size: number of particles in each object
         :rtype: (N1, N3) integers and (N1,) integers"""
         
-        if rank != 0:
-            return None, None
         if obj_type == 'dm':
             # Import hdf5 data
             nb_shs, sh_len, fof_dm_sizes, group_r200, halo_masses = getHDF5SHDMData(self.HDF5_GROUP_DEST, self.WANT_RVIR)
-            print_status(rank,self.start_time,"Gotten getHDF5SHDMData")
-            # Construct catalogue
-            h_cat, h_r200, h_size = calcCSHCat(nb_shs, sh_len, fof_dm_sizes, group_r200, halo_masses, self.MIN_NUMBER_PTCS)
-            print_status(rank,self.start_time,"Gotten calcCSHCat")
-            self.r200 = h_r200
-            del nb_shs; del sh_len; del fof_dm_sizes; del group_r200; del halo_masses; del h_r200
-            return h_cat, h_size
+            if rank == 0:
+                # Construct catalogue
+                h_cat, h_r200, h_size = calcCSHCat(nb_shs, sh_len, fof_dm_sizes, group_r200, halo_masses, self.MIN_NUMBER_PTCS)
+                self.r200 = h_r200
+                del nb_shs; del sh_len; del fof_dm_sizes; del group_r200; del halo_masses
+                return h_cat, h_size
+            else:
+                del nb_shs; del sh_len; del fof_dm_sizes; del group_r200; del halo_masses
+                return None, None
         else:
             if self.r200 is None:
                 # Import hdf5 data
                 nb_shs, sh_len, fof_dm_sizes, group_r200, halo_masses = getHDF5SHDMData(self.HDF5_GROUP_DEST, self.WANT_RVIR)
-                # Construct catalogue
-                h_cat, h_r200, h_pass = calcCSHCat(nb_shs, sh_len, fof_dm_sizes, group_r200, halo_masses, self.MIN_NUMBER_PTCS)
-                self.r200 = h_r200
-                del nb_shs; del sh_len; del fof_dm_sizes; del group_r200; del h_cat; del halo_masses; del h_r200
+                if rank == 0:
+                    # Construct catalogue
+                    h_cat, h_r200, h_pass = calcCSHCat(nb_shs, sh_len, fof_dm_sizes, group_r200, halo_masses, self.MIN_NUMBER_PTCS)
+                    self.r200 = h_r200
+                    del h_cat; del h_pass
+                del nb_shs; del sh_len; del fof_dm_sizes; del group_r200; del halo_masses
             # Import hdf5 data
             nb_shs, sh_len_gx, fof_gx_sizes = getHDF5SHGxData(self.HDF5_GROUP_DEST)
-            # Construct catalogue
-            gx_cat, gx_size = calcGxCat(nb_shs, sh_len_gx, fof_gx_sizes, self.MIN_NUMBER_STAR_PTCS)
-            del nb_shs; del sh_len_gx; del fof_gx_sizes
-            return gx_cat, gx_size
+            if rank == 0:
+                # Construct catalogue
+                gx_cat, gx_size = calcGxCat(nb_shs, sh_len_gx, fof_gx_sizes, self.MIN_NUMBER_STAR_PTCS)
+                del nb_shs; del sh_len_gx; del fof_gx_sizes
+                return gx_cat, gx_size
+            else:
+                del nb_shs; del sh_len_gx; del fof_gx_sizes
+                return None, None
     
     def getMassesCenters(self, list select, str obj_type = 'dm'):
         """ Calculate total mass and centers of objects
[…]
