+++ cosmic_profiles/for_docs/dens_profs/dens_profs_classes.pyx
@@ -154,7 +154,7 @@ cdef class DensProfsHDF5(CosmicBase):
     ``estDensProfs()``, ``fitDensProfs()``, ``estConcentrations()``,
     ``plotDensProfs()``."""
     
-    def __init__(self, str HDF5_SNAP_DEST, str HDF5_GROUP_DEST, str SNAP, float L_BOX, int MIN_NUMBER_PTCS, int MIN_NUMBER_STAR_PTCS, str CENTER, bint WANT_RVIR):
+    def __init__(self, str HDF5_SNAP_DEST, str HDF5_GROUP_DEST, str SNAP, float L_BOX, int MIN_NUMBER_PTCS, str OBJ_TYPE, str CENTER, str RVIR_OR_R200):
         """
         :param HDF5_SNAP_DEST: where we can find the snapshot
         :type HDF5_SNAP_DEST: string
@@ -166,62 +166,66 @@ cdef class DensProfsHDF5(CosmicBase):
         :type SNAP: string
         :param L_BOX: simulation box side length
         :type L_BOX: float, units: Mpc/h
-        :param MIN_NUMBER_PTCS: minimum number of DM particles for halo to qualify for morphology calculation
+        :param MIN_NUMBER_PTCS: minimum number of particles for object to qualify for morphology calculation
         :type MIN_NUMBER_PTCS: int
-        :param MIN_NUMBER_STAR_PTCS: minimum number of star particles for galaxy to qualify for morphology calculation
-        :type MIN_NUMBER_STAR_PTCS: int
         :param CENTER: shape quantities will be calculated with respect to CENTER = 'mode' (point of highest density)
             or 'com' (center of mass) of each halo
         :type CENTER: str
-        :param WANT_RVIR: Whether or not we want quantities (e.g. D_LOGSTART) expressed 
-            with respect to the virial radius R_vir or the overdensity radius R_200
-        :type WANT_RVIR: boolean"""
+        :param RVIR_OR_R200: 'Rvir' if we want quantities (e.g. D_LOGSTART) to be expressed 
+            with respect to the virial radius R_vir, 'R200' for the overdensity radius R_200
+        :type RVIR_OR_R200: str"""
         super().__init__(SNAP, L_BOX, MIN_NUMBER_PTCS, CENTER)
         self.HDF5_SNAP_DEST = HDF5_SNAP_DEST
         self.HDF5_GROUP_DEST = HDF5_GROUP_DEST
-        self.MIN_NUMBER_STAR_PTCS = MIN_NUMBER_STAR_PTCS
-        self.WANT_RVIR = WANT_RVIR
+        self.RVIR_OR_R200 = RVIR_OR_R200
+        self.OBJ_TYPE = OBJ_TYPE
         
-    def getXYZMasses(self, str obj_type = 'dm'):
-        """ Retrieve positions and masses of DM/gx
+    def getPartType(self):
+        """ Return particle type number
         
-        :param obj_type: either 'dm' or 'gx', depending on what catalogue we are looking at
-        :type obj_type: string
-        :return xyz, masses, MIN_NUMBER_PTCS: positions, masses, and minimum number of particles
-        :rtype: (N2,3) floats, (N2,) floats, int"""
+        :returns: particle type number
+        :rtype: int"""
+        return
+    
+    def getXYZMasses(self):
+        """ Retrieve positions and masses of objects
+        
+        :return xyz, masses: positions and masses
+        :rtype: (N2,3) floats, (N2,) floats"""
         return
         
-    def getVelXYZ(self, str obj_type = 'dm'):
-        """ Retrieve velocities of DM/gx
+    def getVelXYZ(self):
+        """ Retrieve velocities of objects
         
[…]
