+++ cosmic_profiles/for_docs/shape_profs/shape_profs_classes.pyx
@@ -253,7 +253,7 @@ cdef class DensShapeProfsHDF5(DensProfsHDF5):
     cdef int IT_WALL
     cdef int IT_MIN
     
-    def __init__(self, str HDF5_SNAP_DEST, str HDF5_GROUP_DEST, str SNAP, float L_BOX, int MIN_NUMBER_PTCS, int MIN_NUMBER_STAR_PTCS, int D_LOGSTART, int D_LOGEND, int D_BINS, float IT_TOL, int IT_WALL, int IT_MIN, str CENTER, bint WANT_RVIR):
+    def __init__(self, str HDF5_SNAP_DEST, str HDF5_GROUP_DEST, str SNAP, float L_BOX, int MIN_NUMBER_PTCS, int D_LOGSTART, int D_LOGEND, int D_BINS, float IT_TOL, int IT_WALL, int IT_MIN, str CENTER, str RVIR_OR_R200):
         """
         :param HDF5_SNAP_DEST: where we can find the snapshot
         :type HDF5_SNAP_DEST: string
@@ -265,10 +265,8 @@ cdef class DensShapeProfsHDF5(DensProfsHDF5):
         :type SNAP: string
         :param L_BOX: simulation box side length
         :type L_BOX: float, units: Mpc/h
-        :param MIN_NUMBER_PTCS: minimum number of DM particles for halo to qualify for morphology calculation
+        :param MIN_NUMBER_PTCS: minimum number of particles for object to qualify for morphology calculation
         :type MIN_NUMBER_PTCS: int
-        :param MIN_NUMBER_STAR_PTCS: minimum number of star particles for galaxy to qualify for morphology calculation
-        :type MIN_NUMBER_STAR_PTCS: int
         :param D_LOGSTART: logarithm of minimum ellipsoidal radius of interest, in units of R200 of parent halo
         :type D_LOGSTART: int
         :param D_LOGEND: logarithm of maximum ellipsoidal radius of interest, in units of R200 of parent halo
@@ -286,10 +284,10 @@ cdef class DensShapeProfsHDF5(DensProfsHDF5):
         :param CENTER: shape quantities will be calculated with respect to CENTER = 'mode' (point of highest density)
             or 'com' (center of mass) of each halo
         :type CENTER: str
-        :param WANT_RVIR: Whether or not we want quantities (e.g. D_LOGSTART) expressed 
-            with respect to the virial radius R_vir or the overdensity radius R_200
-        :type WANT_RVIR: boolean"""
-        super().__init__(HDF5_SNAP_DEST, HDF5_GROUP_DEST, SNAP, L_BOX, MIN_NUMBER_PTCS, MIN_NUMBER_STAR_PTCS, CENTER, WANT_RVIR)
+        :param RVIR_OR_R200: 'Rvir' if we want quantities (e.g. D_LOGSTART) to be expressed 
+            with respect to the virial radius R_vir, 'R200' for the overdensity radius R_200
+        :type RVIR_OR_R200: str"""
+        super().__init__(HDF5_SNAP_DEST, HDF5_GROUP_DEST, SNAP, L_BOX, MIN_NUMBER_PTCS, CENTER, RVIR_OR_R200)
         self.D_LOGSTART = D_LOGSTART
         self.D_LOGEND = D_LOGEND
         self.D_BINS = D_BINS
@@ -297,7 +295,7 @@ cdef class DensShapeProfsHDF5(DensProfsHDF5):
         self.IT_WALL = IT_WALL
         self.IT_MIN = IT_MIN
         
-    def estDensProfs(self, ROverR200, list select, bint direct_binning = True, bint spherical = True, bint reduced = False, bint shell_based = False, str obj_type = 'dm'):
[…]
