### 2022-08-28 25336dd
Fixed getHDF5ObjData() function; updated example_scripts
