+++ cosmic_profiles/gadget_hdf5/get_hdf5.py
@@ -251,7 +251,7 @@ def getHDF5SHData(HDF5_GROUP_DEST, RVIR_OR_R200, PART_TYPE):
         for snap_run in range(rank*perrank, rank*perrank+do_sth*(perrank+last*(nb_jobs_to_do-(rank+1)*perrank))):
             g = h5py.File(r'{}'.format(hdf5GroupFilenamesList[snap_run]), 'r')
             if 'Group/GroupLenType' in g:
-                fof_sizes = np.hstack((fof_sizes, np.int32([np.int32(g['Group/GroupLenType'][i,1]) for i in range(g['Group/GroupLenType'].shape[0])])))
+                fof_sizes = np.hstack((fof_sizes, np.int32([np.int32(g['Group/GroupLenType'][i,PART_TYPE]) for i in range(g['Group/GroupLenType'].shape[0])])))
                 if RVIR_OR_R200 == 'Rvir':
                     group_r200 = np.hstack((group_r200, np.float32(g['Group/Group_R_TopHat200'][:]/1000)))
                 elif RVIR_OR_R200 == 'R200':
@@ -261,7 +261,7 @@ def getHDF5SHData(HDF5_GROUP_DEST, RVIR_OR_R200, PART_TYPE):
                 nb_shs = np.hstack((nb_shs, np.int32([np.int32(g['Group/GroupNsubs'][i]) for i in range(g['Group/GroupNsubs'].shape[0])])))
                 count_fof += g['Group/GroupLenType'].shape[0]
             if 'Subhalo/SubhaloLenType' in g:
-                sh_len = np.hstack((sh_len, np.int32([np.int32(g['Subhalo/SubhaloLenType'][i,1]) for i in range(g['Subhalo/SubhaloLenType'].shape[0])])))        
+                sh_len = np.hstack((sh_len, np.int32([np.int32(g['Subhalo/SubhaloLenType'][i,PART_TYPE]) for i in range(g['Subhalo/SubhaloLenType'].shape[0])])))        
                 count_sh += g['Subhalo/SubhaloLenType'].shape[0]
         
         count_fof_new = comm.gather(count_fof, root=0)
@@ -321,14 +321,17 @@ def getHDF5SHData(HDF5_GROUP_DEST, RVIR_OR_R200, PART_TYPE):
     getHDF5SHData.inner(HDF5_GROUP_DEST, RVIR_OR_R200, PART_TYPE)
     return getHDF5SHData.inner(HDF5_GROUP_DEST, RVIR_OR_R200, PART_TYPE)
 
-def getHDF5DMData(HDF5_SNAP_DEST):
+def getHDF5ObjData(HDF5_SNAP_DEST, PART_TYPE):
     """ Retrieve DM HDF5 data from the simulation box
         
     :param HDF5_SNAP_DEST: path to snapshot, particle data
     :type HDF5_SNAP_DEST: string
-    :return: dm_xyz (DM ptc positions), dm_masses (mass of each DM ptc),
-        dm_velxyz (velocity of each DM ptc)
-    :rtype: float and int arrays"""
+    :param PART_TYPE: which simulation particles to consider, 0 for gas, 1 for DM,
+        4 for stars
+    :type PART_TYPE: int
+    :return: obj_xyz (ptc positions), obj_masses (mass of each ptc),
+        obj_velxyz (velocity of each ptc)
+    :rtype: float arrays"""
     def inner(HDF5_SNAP_DEST):
         dm_x = np.empty(0, dtype = np.float32)
         dm_y = np.empty(0, dtype = np.float32)
@@ -348,14 +351,17 @@ def getHDF5DMData(HDF5_SNAP_DEST):
[…]
