### 2022-08-28 8441d21
Updated docs on OBJ_TYPE etc
