+++ docs/source/shape_estimation/index.rst
@@ -51,7 +51,6 @@ To calculate shape profiles with *Cosmic Profiles*, let us assume we are dealing
     L_BOX = np.float32(10) # cMpc/h
     HDF5_GROUP_DEST = "/path/to/groups_035"
     HDF5_SNAP_DEST = "/path/to/snapdir_035"
-    SNAP_MAX = 16
     D_LOGSTART = -2
     D_LOGEND = 1
     D_BINS = 30 # If D_LOGSTART == -2 D_LOGEND == 1, 60 corresponds to shell width of 0.05 dex
@@ -60,14 +59,25 @@ To calculate shape profiles with *Cosmic Profiles*, let us assume we are dealing
     IT_MIN = 10
     SNAP = '035'
     CENTER = 'mode'
-    MIN_NUMBER_DM_PTCS = 200
-    MIN_NUMBER_STAR_PTCS = 100
-    WANT_RVIR = False # Whether or not we want quantities (e.g. D_LOGSTART) expressed with respect to the virial radius R_vir or overdensity radius R_200
+    MIN_NUMBER_PTCS = 200
+    RVIR_OR_R200 = 'R200'
+    OBJ_TYPE = 'dm' # Which simulation particles to consider, 'dm', 'gas' or 'stars'
 
     # Instantiate object
-    cprofiles = DensShapeProfsHDF5(HDF5_SNAP_DEST, HDF5_GROUP_DEST, SNAP, SNAP_MAX, L_BOX, MIN_NUMBER_DM_PTCS, MIN_NUMBER_STAR_PTCS, D_LOGSTART, D_LOGEND, D_BINS, IT_TOL, IT_WALL, IT_MIN, CENTER, WANT_RVIR)
+    cprofiles = DensShapeProfsHDF5(HDF5_SNAP_DEST, HDF5_GROUP_DEST, SNAP, L_BOX, MIN_NUMBER_PTCS, D_LOGSTART, D_LOGEND, D_BINS, IT_TOL, IT_WALL, IT_MIN, CENTER, RVIR_OR_R200, OBJ_TYPE)
 
-with arguments explained in :ref:`the code reference<Cosmic Profiles Code Reference>`.
+with arguments explained in detail in :ref:`the code reference<Cosmic Profiles Code Reference>` but summarized here for completeness:
+  * ``HDF5_GROUP_DEST`` and ``HDF5_SNAP_DEST``: path to simulation snapshot
+  * ``D_LOGSTART`` and ``D_LOGEND``: logarithm of minimum and maximum ellipsoidal radius of interest, in units of R200 or Rvir (depending on ``RVIR_OR_R200``) of parent halo
+  * ``D_BINS``: number of bins to consider for shape profiling 
+  * ``IT_TOL``: convergence tolerance in shape estimation algorithm, eigenvalue fractions must differ by less than ``IT_TOL`` for algorithm to halt
+  * ``IT_WALL``: maximum permissible number of iterations in shape estimation algorithm
+  * ``IT_MIN``: minimum number of particles (DM, gas or star particles depending on ``OBJ_TYPE``) in any iteration, if undercut, shape is unclassified
+  * ``SNAP``: snapshot identifier, used when dumping files etc.
+  * ``CENTER``: shape quantities will be calculated with respect to CENTER = 'mode' (point of highest density) or 'com' (center of mass) of each object (= DM halo, gas halo or star particle halo)
+  * ``MIN_NUMBER_PTCS``: minimum number of particles for objects to qualify for analyses (e.g. shape analysis)
[…]
