+++ docs/source/conf.py
@@ -32,6 +32,7 @@ release = '1.0.0'
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
+    'sphinx_design', # For dropdown menus
     'sphinx.ext.todo',
     'sphinx.ext.autodoc',
     'sphinx_rtd_theme',
