### 2022-08-28 6bd95cb
Fixed glob.glob bug that resulted in random ordering of snap and group files
