### 2022-08-28 #4626 @tibordome
@adonath Many thanks for editing the submission. We are currently also working on v1.2.7 which will have several improvements and some fixes. We still need to fix the unit confusion that somebody pointed out, should be done in the next few days.
