### 2022-08-29 1a2de9d
Fixed unit system ambiguities by adding updateInUnitSystem() and updateOutUnitSystem
