+++ cosmic_profiles/dens_profs/dens_profs_classes.pyx
@@ -17,24 +17,24 @@ size = comm.Get_size()
 cdef class DensProfs(CosmicBase):
     """ Class for density profile calculations
     
-    Its public methods are ``getIdxCat()``,
-    ``getMassesCenters()``, ``estDensProfs``, ``fitDensProfs()``, ``estConcentrations()``, 
-    ``plotDensProfs()``."""
+    Its public methods are ``getR200s()``, ``getIdxCat()``,
+    ``getXYZMasses()``, ``getMassesCenters()``, ``_getMassesCenters()``, ``estDensProfs()``, 
+    ``fitDensProfs()``, ``estConcentrations()``, ``plotDensProfs()``."""
     
     def __init__(self, float[:,:] xyz, float[:] masses, idx_cat, float[:] r200, str SNAP, float L_BOX, int MIN_NUMBER_PTCS, str CENTER):
         """
-        :param xyz: positions of all simulation particles
+        :param xyz: positions of all simulation particles in config.InUnitLength_in_cm
         :type xyz: (N2,3) floats, N2 >> N1
-        :param masses: masses of all simulation particles
+        :param masses: masses of all simulation particles in config.InUnitMass_in_g
         :type masses: (N2,) floats
         :param idx_cat: each entry of the list is a list containing indices of particles belonging to an object
         :type idx_cat: list of length N1
-        :param r200: R_200 radii of the parent halos
+        :param r200: R_200 radii of the parent halos in config.InUnitLength_in_cm
         :type r200: (N1,) floats
         :param SNAP: snapshot identifier, e.g. '024'
         :type SNAP: string
-        :param L_BOX: simulation box side length
-        :type L_BOX: float, units: Mpc/h
+        :param L_BOX: simulation box side length in config.InUnitLength_in_cm
+        :type L_BOX: float
         :param MIN_NUMBER_PTCS: minimum number of particles for object to qualify for morphology calculation
         :type MIN_NUMBER_PTCS: int
         :param CENTER: shape quantities will be calculated with respect to CENTER = 'mode' (point of highest density)
@@ -42,8 +42,8 @@ cdef class DensProfs(CosmicBase):
         :type CENTER: str"""
         super().__init__(SNAP, L_BOX, MIN_NUMBER_PTCS, CENTER)
         assert xyz.shape[0] == masses.shape[0], "xyz.shape[0] must be equal to masses.shape[0]"
-        self.xyz = xyz.base
-        self.masses = masses.base
+        self.xyz = xyz.base*InUnitLength_in_cm/3.085678e24 # self.xyz will be in Mpc/h
+        self.masses = masses.base*InUnitMass_in_g/1.989e43 # self.masses will be in 10^10 M_sun/h
         cdef int nb_objs = len(idx_cat)
         cdef int p
         cdef int[:] obj_pass = np.zeros((nb_objs,), dtype = np.int32)
@@ -61,9 +61,18 @@ cdef class DensProfs(CosmicBase):
                 cat_arr.base[idxs_compr[p],:obj_size[p]] = np.array(idx_cat[p])
         self.idx_cat = cat_arr.base
         self.obj_size = obj_size.base[obj_pass.base.nonzero()[0]]
-        self.r200 = r200.base[obj_pass.base.nonzero()[0]]
+        self.r200 = r200.base[obj_pass.base.nonzero()[0]]*InUnitLength_in_cm/3.085678e24 # self.r200 will be in Mpc/h
        
-    def getIdxCat(self):
+    def getR200s(self): # Public Method
+        """ Get overdensity radii in config.OutUnitLength_in_cm units"""
+        print_status(rank,self.start_time,'Starting getR200s() with snap {0}'.format(self.SNAP))
+        
[…]
