+++ cosmic_profiles/common/cosmic_base_class.pyx
@@ -10,6 +10,7 @@ from cosmic_profiles.dens_profs.dens_profs_tools import fitDensProfHelper
 from cosmic_profiles.shape_profs.shape_profs_algos import calcMorphLocal, calcMorphGlobal, calcMorphLocalVelDisp, calcMorphGlobalVelDisp
 from cosmic_profiles.dens_profs.dens_profs_algos import calcMassesCenters, calcDensProfsSphDirectBinning, calcDensProfsEllDirectBinning, calcDensProfsKernelBased
 import time
+import config
 from mpi4py import MPI
 comm = MPI.COMM_WORLD
 rank = comm.Get_rank()
@@ -19,67 +20,58 @@ size = comm.Get_size()
 cdef class CosmicBase:
     """ Parent class governing high-level cosmic shape calculations
     
-    Its public methods are ``getR200s()``, ``getMassesCentersBase()``, 
-    ``getShapeCatLocalBase()``, ``getShapeCatGlobalBase()``, ``getShapeCatVelLocalBase()``, 
-    ``getShapeCatVelGlobalBase()``, ``dumpShapeCatLocalBase()``, ``dumpShapeCatGlobalBase()``,
-    ``dumpShapeCatVelLocalBase()``, ``dumpShapeCatVelGlobalBase()``, ``plotShapeProfsBase()``,
-    ``plotLocalTHistBase()``, ``plotGlobalTHistBase()``, ``getDensProfsBestFitsBase()``,
-    ``getConcentrationsBase()``, ``getDensProfsSphDirectBinningBase()``, ``getDensProfsEllDirectBinningBase()``,
-    ``getDensProfsKernelBasedBase()``, ``getObjInfoBase()``"""
+    Its public methods are ``_getMassesCentersBase()``, 
+    ``_getShapeCatLocalBase()``, ``_getShapeCatGlobalBase()``, ``_getShapeCatVelLocalBase()``, 
+    ``_getShapeCatVelGlobalBase()``, ``_dumpShapeCatLocalBase()``, ``_dumpShapeCatGlobalBase()``,
+    ``_dumpShapeCatVelLocalBase()``, ``_dumpShapeCatVelGlobalBase()``, ``_plotShapeProfsBase()``,
+    ``_plotLocalTHistBase()``, ``_plotGlobalTHistBase()``, ``_getDensProfsBestFitsBase()``,
+    ``_getConcentrationsBase()``, ``_getDensProfsSphDirectBinningBase()``, ``_getDensProfsEllDirectBinningBase()``,
+    ``_getDensProfsKernelBasedBase()``, ``_getObjInfoBase()``"""
     
     def __init__(self, str SNAP, float L_BOX, int MIN_NUMBER_PTCS, str CENTER):
         """
         :param SNAP: snapshot identifier, e.g. '024'
         :type SNAP: string
-        :param L_BOX: simulation box side length
-        :type L_BOX: float, units: Mpc/h
+        :param L_BOX: simulation box side length in config.InUnitLength_in_cm
+        :type L_BOX: float
         :param MIN_NUMBER_PTCS: minimum number of particles for object to qualify for morphology calculation
         :type MIN_NUMBER_PTCS: int
         :param CENTER: shape quantities will be calculated with respect to CENTER = 'mode' (point of highest density)
             or 'com' (center of mass) of each halo
         :type CENTER: str"""
         self.SNAP = SNAP
-        self.L_BOX = L_BOX
+        self.L_BOX = L_BOX*InUnitLength_in_cm/3.085678e24 # self.L_BOX will be in Mpc/h
         self.CENTER = CENTER
         self.MIN_NUMBER_PTCS = MIN_NUMBER_PTCS
         self.start_time = time.time()
         self.SAFE = 6
[…]
