+++ example_scripts/gadget3_hdf5.py
@@ -14,10 +14,12 @@ import inspect
 import numpy as np
 currentdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
 sys.path.append(os.path.join(currentdir, '..')) # Only needed if cosmic_profiles is not installed
-from cosmic_profiles import DensShapeProfsHDF5
+from cosmic_profiles import DensShapeProfsHDF5, updateInUnitSystem, updateOutUnitSystem
 
 # Parameters
-L_BOX = np.float32(10) # Mpc/h
+updateInUnitSystem(in_unit_length_in_cm = 3.085678e21, in_unit_mass_in_g = 1.989e43, in_unit_velocity_in_cm_per_s = 1e5)
+updateOutUnitSystem(in_unit_length_in_cm = 3.085678e24, in_unit_mass_in_g = 1.989e33, in_unit_velocity_in_cm_per_s = 1e5)
+L_BOX = np.float32(10000) # kpc/h
 CAT_DEST = "./cat"
 HDF5_GROUP_DEST = "../../code_git/example_snapshot/LLGas256b20/groups_035"
 HDF5_SNAP_DEST = "../../code_git/example_snapshot/LLGas256b20/snapdir_035"

[…]
