+++ cosmic_profiles/shape_profs/shape_profs_classes.pyx
@@ -16,6 +16,7 @@ from cosmic_profiles.common.python_routines import print_status, set_axes_equal,
 from cosmic_profiles.shape_profs.shape_profs_tools import getGlobalEpsHist, getLocalEpsHist
 from cosmic_profiles.gadget_hdf5.get_hdf5 import getHDF5SHData
 import time
+import config
 from mpi4py import MPI
 comm = MPI.COMM_WORLD
 rank = comm.Get_rank()
@@ -109,7 +110,7 @@ cdef class DensShapeProfs(DensProfs):
                     dens_profs = self._getDensProfsSphDirectBinningBase(self.xyz.base, self.masses.base, self.r200.base[select[0]:select[1]+1], self.idx_cat.base[select[0]:select[1]+1], self.obj_size.base[select[0]:select[1]+1], np.float32(ROverR200))
             else:
                 dens_profs = self._getDensProfsKernelBasedBase(self.xyz.base, self.masses.base, self.r200.base[select[0]:select[1]+1], self.idx_cat.base[select[0]:select[1]+1], self.obj_size.base[select[0]:select[1]+1], np.float32(ROverR200))
-            return dens_profs*1.989e33/OutUnitMass_in_g*(3.085678e24/OutUnitLength_in_cm)**(-3)
+            return dens_profs*1.989e33/config.OutUnitMass_in_g*(3.085678e24/config.OutUnitLength_in_cm)**(-3)
         else:
             return None
         
@@ -130,7 +131,7 @@ cdef class DensShapeProfs(DensProfs):
         print_status(rank,self.start_time,'Starting getShapeCatLocal() with snap {0}'.format(self.SNAP))
         if rank == 0:
             d, q, s, minor, inter, major, obj_centers, obj_masses = self._getShapeCatLocalBase(self.xyz.base, self.masses.base, self.r200.base[select[0]:select[1]+1], self.idx_cat.base[select[0]:select[1]+1], self.obj_size.base[select[0]:select[1]+1], self.D_LOGSTART, self.D_LOGEND, self.D_BINS, self.IT_TOL, self.IT_WALL, self.IT_MIN, reduced, shell_based)
-            return d, q, s, minor, inter, major, obj_centers*3.085678e24/OutUnitLength_in_cm, obj_masses*self.MASS_UNIT*1.989e33/OutUnitMass_in_g
+            return d, q, s, minor, inter, major, obj_centers*3.085678e24/config.OutUnitLength_in_cm, obj_masses*self.MASS_UNIT*1.989e33/config.OutUnitMass_in_g
         else:
             return None, None, None, None, None, None, None, None
     
@@ -149,7 +150,7 @@ cdef class DensShapeProfs(DensProfs):
         print_status(rank,self.start_time,'Starting getShapeCatGlobal() with snap {0}'.format(self.SNAP))
         if rank == 0:
             d, q, s, minor, inter, major, obj_centers, obj_masses = self._getShapeCatGlobalBase(self.xyz.base, self.masses.base, self.r200.base[select[0]:select[1]+1], self.idx_cat.base[select[0]:select[1]+1], self.obj_size.base[select[0]:select[1]+1], self.IT_TOL, self.IT_WALL, self.IT_MIN, reduced)
-            return d, q, s, minor, inter, major, obj_centers*3.085678e24/OutUnitLength_in_cm, obj_masses*self.MASS_UNIT*1.989e33/OutUnitMass_in_g
+            return d, q, s, minor, inter, major, obj_centers*3.085678e24/config.OutUnitLength_in_cm, obj_masses*self.MASS_UNIT*1.989e33/config.OutUnitMass_in_g
         else:
             return None, None, None, None, None, None, None, None
     
[…]
