+++ cosmic_profiles/tests/test_shapes_ex_script.py
@@ -21,7 +21,7 @@ def test_shapes_ex_script():
     
     #################################### Parameters ################################################
     updateInUnitSystem(in_unit_length_in_cm = 3.085678e24, in_unit_mass_in_g = 1.989e33, in_unit_velocity_in_cm_per_s = 1e5)
-    updateOutUnitSystem(in_unit_length_in_cm = 3.085678e24, in_unit_mass_in_g = 1.989e33, in_unit_velocity_in_cm_per_s = 1e5)
+    updateOutUnitSystem(out_unit_length_in_cm = 3.085678e24, out_unit_mass_in_g = 1.989e33, out_unit_velocity_in_cm_per_s = 1e5)
     L_BOX = np.float32(10) # Mpc/h
     CAT_DEST = "./cosmic_profiles/tests/cat"
     VIZ_DEST = "./cosmic_profiles/tests/viz"
