### 2022-08-29 1fce62b
Added missing public methods of DensShapeProfsHDF5 and DensShapeProfs to docs
