+++ cosmic_profiles/for_docs/shape_profs/shape_profs_classes.c
@@ -1034,7 +1034,7 @@ struct __pyx_obj_11shape_profs_19shape_profs_classes_DensShapeProfs {
 };
 
 
-/* "shape_profs/shape_profs_classes.pyx":240
+/* "shape_profs/shape_profs_classes.pyx":241
  *         return
  * 
  * cdef class DensShapeProfsHDF5(DensProfsHDF5):             # <<<<<<<<<<<<<<
@@ -1969,6 +1969,7 @@ static const char __pyx_k_D_LOGEND[] = "D_LOGEND";
 static const char __pyx_k_Ellipsis[] = "Ellipsis";
 static const char __pyx_k_Get_rank[] = "Get_rank";
 static const char __pyx_k_Get_size[] = "Get_size";
+static const char __pyx_k_OBJ_TYPE[] = "OBJ_TYPE";
 static const char __pyx_k_VIZ_DEST[] = "VIZ_DEST";
 static const char __pyx_k_getstate[] = "__getstate__";
 static const char __pyx_k_itemsize[] = "itemsize";
@@ -2077,6 +2078,7 @@ static PyObject *__pyx_n_s_MemoryError;
 static PyObject *__pyx_kp_s_MemoryView_of_r_at_0x_x;
 static PyObject *__pyx_kp_s_MemoryView_of_r_object;
 static PyObject *__pyx_n_b_O;
+static PyObject *__pyx_n_s_OBJ_TYPE;
 static PyObject *__pyx_kp_s_Out_of_bounds_on_buffer_access_a;
 static PyObject *__pyx_n_s_PickleError;
 static PyObject *__pyx_n_s_ROverR200;
@@ -2186,7 +2188,7 @@ static PyObject *__pyx_pf_11shape_profs_19shape_profs_classes_14DensShapeProfs_2
 static PyObject *__pyx_pf_11shape_profs_19shape_profs_classes_14DensShapeProfs_26getObjInfo(CYTHON_UNUSED struct __pyx_obj_11shape_profs_19shape_profs_classes_DensShapeProfs *__pyx_v_self); /* proto */
 static PyObject *__pyx_pf_11shape_profs_19shape_profs_classes_14DensShapeProfs_28__reduce_cython__(struct __pyx_obj_11shape_profs_19shape_profs_classes_DensShapeProfs *__pyx_v_self); /* proto */
 static PyObject *__pyx_pf_11shape_profs_19shape_profs_classes_14DensShapeProfs_30__setstate_cython__(struct __pyx_obj_11shape_profs_19shape_profs_classes_DensShapeProfs *__pyx_v_self, PyObject *__pyx_v___pyx_state); /* proto */
-static int __pyx_pf_11shape_profs_19shape_profs_classes_18DensShapeProfsHDF5___init__(struct __pyx_obj_11shape_profs_19shape_profs_classes_DensShapeProfsHDF5 *__pyx_v_self, PyObject *__pyx_v_HDF5_SNAP_DEST, PyObject *__pyx_v_HDF5_GROUP_DEST, PyObject *__pyx_v_SNAP, float __pyx_v_L_BOX, int __pyx_v_MIN_NUMBER_PTCS, int __pyx_v_D_LOGSTART, int __pyx_v_D_LOGEND, int __pyx_v_D_BINS, float __pyx_v_IT_TOL, int __pyx_v_IT_WALL, int __pyx_v_IT_MIN, PyObject *__pyx_v_CENTER, PyObject *__pyx_v_RVIR_OR_R200); /* proto */
[…]
