+++ cosmic_profiles/__init__.py
@@ -1,6 +1,8 @@
 from .common.config import (
     updateCachingMaxGBs,
-    updateCachingMaxSize
+    updateCachingMaxSize,
+    updateInUnitSystem,
+    updateOutUnitSystem
 )
 
 from .common.python_routines import (
