### 2022-08-29 92ca9f7
Fixed GB vs gbs bug in test_densities_ex_script.py
