### 2022-08-29 b49550d
Fixed ellipsoidal radii units; Fixed self.masses etc definition to respect float32 data type
