+++ cosmic_profiles/common/cosmic_base_class.pyx
@@ -287,6 +287,7 @@ cdef class CosmicBase:
         """
         if rank == 0:
             d, q, s, minor, inter, major, obj_centers, obj_masses = self._getShapeCatLocalBase(xyz.base, masses.base, r200.base, idx_cat.base, obj_size.base, D_LOGSTART, D_LOGEND, D_BINS, IT_TOL, IT_WALL, IT_MIN, reduced, shell_based)
+            d = d*3.085678e24/config.OutUnitLength_in_cm
             obj_centers = obj_centers*3.085678e24/config.OutUnitLength_in_cm
             obj_masses = obj_masses*self.MASS_UNIT*1.989e33/config.OutUnitMass_in_g
             if d.shape[0] != 0:
@@ -337,6 +338,7 @@ cdef class CosmicBase:
         :type reduced: boolean"""
         if rank == 0:
             d, q, s, minor, inter, major, obj_centers, obj_masses = self._getShapeCatGlobalBase(xyz.base, masses.base, r200.base, idx_cat.base, obj_size.base, IT_TOL, IT_WALL, IT_MIN, reduced)
+            d = d*3.085678e24/config.OutUnitLength_in_cm
             obj_centers = obj_centers*3.085678e24/config.OutUnitLength_in_cm
             obj_masses = obj_masses*self.MASS_UNIT*1.989e33/config.OutUnitMass_in_g
             if d.shape[0] != 0:
@@ -398,6 +400,7 @@ cdef class CosmicBase:
         """
         if rank == 0:
             d, q, s, minor, inter, major, obj_centers, obj_masses = self._getShapeCatVelLocalBase(xyz.base, velxyz.base, masses.base, r200.base, idx_cat.base, obj_size.base, D_LOGSTART, D_LOGEND, D_BINS, IT_TOL, IT_WALL, IT_MIN, reduced, shell_based)
+            d = d*3.085678e24/config.OutUnitLength_in_cm
             obj_centers = obj_centers*3.085678e24/config.OutUnitLength_in_cm
             obj_masses = obj_masses*self.MASS_UNIT*1.989e33/config.OutUnitMass_in_g
             if d.shape[0] != 0:
@@ -458,6 +461,7 @@ cdef class CosmicBase:
         
         if rank == 0:
             d, q, s, minor, inter, major, obj_centers, obj_masses = self._getShapeCatVelGlobalBase(xyz.base, velxyz.base, masses.base, r200.base, idx_cat.base, obj_size.base, IT_TOL, IT_WALL, IT_MIN, reduced)
+            d = d*3.085678e24/config.OutUnitLength_in_cm
             obj_centers = obj_centers*3.085678e24/config.OutUnitLength_in_cm
             obj_masses = obj_masses*self.MASS_UNIT*1.989e33/config.OutUnitMass_in_g
             if d.shape[0] != 0:
