### 2022-08-30 8dce245
Fixed gadget3_hdf5.py incompatibilities; fixed ordering of multiplications with large unit-converting numbers to avoid infinities; updated docs, added warnings
