### 2022-08-30 d3eda6c
Added .base to _getShapeCatGlobalBase() in vizGlobalShapes() etc
