+++ setup.py
@@ -82,7 +82,7 @@ ext_modules=[Extension(
 # This call to setup() does all the work
 setup(
     name="cosmic_profiles",
-    version="1.2.7",
+    version="1.2.6",
     description="Implements 3D point cloud algorithms for estimation and fitting of shape and density profiles",
     long_description=README,
     long_description_content_type="text/x-rst",
