+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 1.2.7
+current_version = 1.2.6
 
 [build_sphinx]
 source-dir = docs/source
