### 2022-08-30 04c282d
Added sphinx_design to docs/environment.yml
