+++ docs/environment.yml
@@ -14,4 +14,5 @@ dependencies:
   - psutil
   - sphinx-autoapi
   - sphinx_rtd_theme
+  - sphinx_design
   - twine
