### 2022-08-30 6bbaa84
Added sphinx_design to docs/environment.yml
