+++ docs/environment.yml
@@ -14,5 +14,6 @@ dependencies:
   - psutil
   - sphinx-autoapi
   - sphinx_rtd_theme
-  - sphinx_design
   - twine
+  - pip:
+    - sphinx_design
