### 2022-08-31 #4626 @tibordome
Thank you @adonath for your patience. We have published the new version v1.2.7. In regards to reviewers we could suggest Sownak Bose or Benedikt Diemer, though we are not sure how busy they are.
