### 2022-11-14 #4626 @adonath
Thanks @tibordome! I had still no luck with the 2nd reviewer, so far only rejections. However I'm still trying...If I still don't get any positive answer soon, I think the JOSS rules might allow to assign myself as 2nd reviewer. I'll double check and keep you up to date.
