### 2022-12-13 #5008 @editorialbot
[REVIEW]: CosmicProfiles: A Python package for radial profiling of finitely sampled dark matter halos and galaxies
**Submitting author:** @tibordome (<a href="http://orcid.org/0000-0003-2586-3702">Tibor Dome</a>)
**Repository:** https://github.com/tibordome/cosmic_profiles
**Branch with paper.md** (empty if default branch): dev
**Version:** v1.4.0
