### 2022-12-30 tibordome/cosmic_profiles#1 @benediktdiemer
Installation via pip fails
I was trying to install the code to begin my review, but the pip installation fails on my laptop (2019 MacBook Pro). The full console output is below. The key error seems to be

clang: error: unsupported option '-fopenmp'

The installation section of the docs does not specify whether OpenMP is actually needed for cosmic-profiles? If so, some additional instructions would be welcome. Building with OpenMP is notoriously difficult on some compilers/systems, namely on Macs.

`Benedikts-MacBook:~ benedito$ sudo pip install cosmic-profiles
dbus[8926]: Dynamic session lookup supported but failed: launchd did not provide a socket path, verify that org.freedesktop.dbus-session.plist is loaded!
dbus[8926]: Dynamic session lookup supported but failed: launchd did not provide a socket path, verify that org.freedesktop.dbus-session.plist is loaded!
WARNING: The directory '/Users/benedito/Library/Caches/pip' or its parent directory is not owned or is not writable by the current user. The cache has been disabled. Check the permissions and owner of that directory. If executing pip with sudo, you should use sudo's -H flag.
Collecting cosmic-profiles
  Downloading cosmic_profiles-1.2.7.tar.gz (1.4 MB)
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Requirement already satisfied: matplotlib in /opt/local/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages (from cosmic-profiles) (3.3.3)
Collecting pathos
  Downloading pathos-0.3.0-py3-none-any.whl (79 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 79.8/79.8 kB 113.5 MB/s eta 0:00:00
Requirement already satisfied: cython in /opt/local/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages (from cosmic-profiles) (0.29.24)
Collecting mpi4py
  Downloading mpi4py-3.1.4.tar.gz (2.5 MB)
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Requirement already satisfied: scipy in /opt/local/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages (from cosmic-profiles) (1.7.3)
Collecting cythongsl
  Downloading CythonGSL-0.2.2.tar.gz (85 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 85.4/85.4 kB 109.2 MB/s eta 0:00:00
  Preparing metadata (setup.py) ... done
Requirement already satisfied: scikit-learn in /opt/local/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages (from cosmic-profiles) (1.0.1)
Requirement already satisfied: h5py in /opt/local/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages (from cosmic-profiles) (3.1.0)
Requirement already satisfied: numpy>=1.19.2 in /opt/local/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages (from cosmic-profiles) (1.21.5)
Collecting psutil
  Downloading psutil-5.9.4-cp36-abi3-macosx_10_9_x86_64.whl (243 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 243.5/243.5 kB 24.0 MB/s eta 0:00:00
Requirement already satisfied: cached-property in /opt/local/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages (from h5py->cosmic-profiles) (1.5.2)
Requirement already satisfied: cycler>=0.10 in /opt/local/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages (from matplotlib->cosmic-profiles) (0.10.0)
Requirement already satisfied: kiwisolver>=1.0.1 in /opt/local/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages (from matplotlib->cosmic-profiles) (1.3.1)
Requirement already satisfied: pillow>=6.2.0 in /opt/local/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages (from matplotlib->cosmic-profiles) (8.1.0)
Requirement already satisfied: pyparsing!=2.0.4,!=2.1.2,!=2.1.6,>=2.0.3 in /opt/local/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages (from matplotlib->cosmic-profiles) (3.0.9)
Requirement already satisfied: python-dateutil>=2.1 in /opt/local/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages (from matplotlib->cosmic-profiles) (2.8.2)
Collecting ppft>=1.7.6.6
  Downloading ppft-1.7.6.6-py3-none-any.whl (52 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 52.8/52.8 kB 94.2 MB/s eta 0:00:00
Collecting pox>=0.3.2
  Downloading pox-0.3.2-py3-none-any.whl (29 kB)
Collecting multiprocess>=0.70.14
  Downloading multiprocess-0.70.14-py37-none-any.whl (115 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 115.7/115.7 kB 110.2 MB/s eta 0:00:00
Collecting dill>=0.3.6
  Downloading dill-0.3.6-py3-none-any.whl (110 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 110.5/110.5 kB 3.1 MB/s eta 0:00:00
Requirement already satisfied: joblib>=0.11 in /opt/local/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages (from scikit-learn->cosmic-profiles) (1.1.0)
Requirement already satisfied: threadpoolctl>=2.0.0 in /opt/local/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages (from scikit-learn->cosmic-profiles) (3.0.0)
Requirement already satisfied: six in /opt/local/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages (from cycler>=0.10->matplotlib->cosmic-profiles) (1.16.0)
Building wheels for collected packages: cosmic-profiles, cythongsl, mpi4py
  Building wheel for cosmic-profiles (pyproject.toml) ... error
  error: subprocess-exited-with-error

  × Building wheel for cosmic-profiles (pyproject.toml) did not run successfully.
  │ exit code: 1
  ╰─> [102 lines of output]
      running bdist_wheel
      running build
      running build_py
      creating build
      creating build/lib.macosx-10.15-x86_64-cpython-37
      creating build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles
      copying cosmic_profiles/__init__.py -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles
      creating build/lib.macosx-10.15-x86_64-cpython-37/example_scripts
      copying example_scripts/gadget3_hdf5.py -> build/lib.macosx-10.15-x86_64-cpython-37/example_scripts
      copying example_scripts/apply_ahf.py -> build/lib.macosx-10.15-x86_64-cpython-37/example_scripts
      creating build/lib.macosx-10.15-x86_64-cpython-37/docs
      creating build/lib.macosx-10.15-x86_64-cpython-37/docs/source
      copying docs/source/conf.py -> build/lib.macosx-10.15-x86_64-cpython-37/docs/source
      creating build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/dens_profs
      copying cosmic_profiles/dens_profs/dens_profs_tools.py -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/dens_profs
      creating build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/tests
      copying cosmic_profiles/tests/test_shapes.py -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/tests
      copying cosmic_profiles/tests/test_densities.py -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/tests
      copying cosmic_profiles/tests/test_shapes_ex_script.py -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/tests
      copying cosmic_profiles/tests/test_densities_ex_script.py -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/tests
      creating build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/mock_tools
      copying cosmic_profiles/mock_tools/mock_halo_gen.py -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/mock_tools
      creating build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/shape_profs
      copying cosmic_profiles/shape_profs/shape_profs_tools.py -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/shape_profs
      creating build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/gadget_hdf5
      copying cosmic_profiles/gadget_hdf5/get_hdf5.py -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/gadget_hdf5
      creating build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/common
      copying cosmic_profiles/common/config.py -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/common
      copying cosmic_profiles/common/python_routines.py -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/common
      copying cosmic_profiles/common/caching.py -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/common
      copying cosmic_profiles/common/cosmo_tools.py -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/common
      creating build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs
      copying cosmic_profiles/for_docs/setup_compile.py -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs
      creating build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/dens_profs
      copying cosmic_profiles/for_docs/dens_profs/dens_profs_tools.py -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/dens_profs
      creating build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/mock_tools
      copying cosmic_profiles/for_docs/mock_tools/mock_uni.py -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/mock_tools
      copying cosmic_profiles/for_docs/mock_tools/mock_halo_gen.py -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/mock_tools
      creating build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/shape_profs
      copying cosmic_profiles/for_docs/shape_profs/shape_profs_tools.py -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/shape_profs
      creating build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/gadget_hdf5
      copying cosmic_profiles/for_docs/gadget_hdf5/get_hdf5.py -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/gadget_hdf5
      creating build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/common
      copying cosmic_profiles/for_docs/common/config.py -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/common
      copying cosmic_profiles/for_docs/common/python_routines.py -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/common
      copying cosmic_profiles/for_docs/common/__init__.py -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/common
      copying cosmic_profiles/for_docs/common/caching.py -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/common
      copying cosmic_profiles/for_docs/common/cosmo_tools.py -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/common
      running egg_info
      writing cosmic_profiles.egg-info/PKG-INFO
      writing dependency_links to cosmic_profiles.egg-info/dependency_links.txt
      writing requirements to cosmic_profiles.egg-info/requires.txt
      writing top-level names to cosmic_profiles.egg-info/top_level.txt
      reading manifest file 'cosmic_profiles.egg-info/SOURCES.txt'
      reading manifest template 'MANIFEST.in'
      adding license file 'LICENSE.md'
      writing manifest file 'cosmic_profiles.egg-info/SOURCES.txt'
      copying cosmic_profiles/__init__.pxd -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles
      creating build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/cython_helpers
      copying cosmic_profiles/cython_helpers/__init__.pxd -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/cython_helpers
      copying cosmic_profiles/cython_helpers/helper_class.c -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/cython_helpers
      copying cosmic_profiles/cython_helpers/helper_class.pxd -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/cython_helpers
      copying cosmic_profiles/cython_helpers/helper_class.pyx -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/cython_helpers
      copying cosmic_profiles/dens_profs/__init__.pxd -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/dens_profs
      copying cosmic_profiles/dens_profs/dens_profs_algos.c -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/dens_profs
      copying cosmic_profiles/dens_profs/dens_profs_algos.pyx -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/dens_profs
      copying cosmic_profiles/dens_profs/dens_profs_classes.c -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/dens_profs
      copying cosmic_profiles/dens_profs/dens_profs_classes.pxd -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/dens_profs
      copying cosmic_profiles/dens_profs/dens_profs_classes.pyx -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/dens_profs
      copying cosmic_profiles/shape_profs/__init__.pxd -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/shape_profs
      copying cosmic_profiles/shape_profs/shape_profs_algos.c -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/shape_profs
      copying cosmic_profiles/shape_profs/shape_profs_algos.pxd -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/shape_profs
      copying cosmic_profiles/shape_profs/shape_profs_algos.pyx -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/shape_profs
      copying cosmic_profiles/shape_profs/shape_profs_classes.c -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/shape_profs
      copying cosmic_profiles/shape_profs/shape_profs_classes.pyx -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/shape_profs
      copying cosmic_profiles/gadget_hdf5/__init__.pxd -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/gadget_hdf5
      copying cosmic_profiles/gadget_hdf5/gen_catalogues.c -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/gadget_hdf5
      copying cosmic_profiles/gadget_hdf5/gen_catalogues.pyx -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/gadget_hdf5
      copying cosmic_profiles/common/__init__.pxd -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/common
      copying cosmic_profiles/common/cosmic_base_class.c -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/common
      copying cosmic_profiles/common/cosmic_base_class.pxd -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/common
      copying cosmic_profiles/common/cosmic_base_class.pyx -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/common
      creating build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/cython_helpers
      copying cosmic_profiles/for_docs/cython_helpers/helper_class.pyx -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/cython_helpers
      copying cosmic_profiles/for_docs/dens_profs/__init__.pxd -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/dens_profs
      copying cosmic_profiles/for_docs/dens_profs/dens_profs_algos.pyx -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/dens_profs
      copying cosmic_profiles/for_docs/dens_profs/dens_profs_classes.pxd -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/dens_profs
      copying cosmic_profiles/for_docs/dens_profs/dens_profs_classes.pyx -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/dens_profs
      copying cosmic_profiles/for_docs/shape_profs/shape_profs_algos.pyx -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/shape_profs
      copying cosmic_profiles/for_docs/shape_profs/shape_profs_classes.pyx -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/shape_profs
      copying cosmic_profiles/for_docs/gadget_hdf5/gen_catalogues.pyx -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/gadget_hdf5
      copying cosmic_profiles/for_docs/common/__init__.pxd -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/common
      copying cosmic_profiles/for_docs/common/cosmic_base_class.pxd -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/common
      copying cosmic_profiles/for_docs/common/cosmic_base_class.pyx -> build/lib.macosx-10.15-x86_64-cpython-37/cosmic_profiles/for_docs/common
      running build_ext
      building 'cosmic_profiles.cython_helpers.helper_class' extension
      creating build/temp.macosx-10.15-x86_64-cpython-37
      creating build/temp.macosx-10.15-x86_64-cpython-37/cosmic_profiles
      creating build/temp.macosx-10.15-x86_64-cpython-37/cosmic_profiles/cython_helpers
      /usr/bin/clang -Wno-unused-result -Wsign-compare -Wunreachable-code -fno-common -dynamic -DNDEBUG -g -fwrapv -O3 -Wall -pipe -I/opt/local/libexec/openssl11/include -Os -isysroot/Library/Developer/CommandLineTools/SDKs/MacOSX10.15.sdk -I/private/tmp/pip-build-env-cawu8_a5/overlay/lib/python3.7/site-packages/numpy/core/include -I. -I/opt/local/Library/Frameworks/Python.framework/Versions/3.7/include/python3.7m -c cosmic_profiles/cython_helpers/helper_class.c -o build/temp.macosx-10.15-x86_64-cpython-37/cosmic_profiles/cython_helpers/helper_class.o -fopenmp
      clang: error: unsupported option '-fopenmp'
      error: command '/usr/bin/clang' failed with exit code 1
      [end of output]

  ERROR: Failed building wheel for cosmic-profiles
  Building wheel for cythongsl (setup.py) ... done
  Created wheel for cythongsl: filename=CythonGSL-0.2.2-py3-none-any.whl size=44179 sha256=bc4647fdd37e895904250af045df418261fa897994d9aefec7a80d4a7cd260ac
  Stored in directory: /private/tmp/pip-ephem-wheel-cache-l3i_nr7l/wheels/b9/78/db/06ed1a1942089f436eec41cb761c99ba601a421cca9eb91915
  Building wheel for mpi4py (pyproject.toml) ... done
  Created wheel for mpi4py: filename=mpi4py-3.1.4-cp37-cp37m-macosx_10_15_x86_64.whl size=490145 sha256=0e7c526e60a2d9c4cb567ae2f0aa701cbf49f20a66271fc673b9dce7853e80d5
  Stored in directory: /private/tmp/pip-ephem-wheel-cache-l3i_nr7l/wheels/87/81/66/c445deac61f7fb9b727c7b87ecc42dba6b948a03674b2fc56d
Successfully built cythongsl mpi4py
Failed to build cosmic-profiles
ERROR: Could not build wheels for cosmic-profiles, which is required to install pyproject.toml-based projects
Benedikts-MacBook:~ benedito$ `
