### 2022-12-30 #5008 @benediktdiemer
Just added issue https://github.com/tibordome/cosmic_profiles/issues/1 (cannot build cosmic-profiles on my computer).
