### 2023-01-07 tibordome/cosmic_profiles#1 @tibordome
@benediktdiemer Apologies for the delay.

I haven't checked the installation myself on mac since the conda forge builds were successful and thus seemed ok. I think the conda install automatically takes care of the openmp dependency whereas the pip install might assume that openmp is available (via the -fopenmp flag of clang). Installing openmp on mac is possible (but difficult indeed) according to https://stackoverflow.com/questions/35134681/installing-openmp-on-mac-os-x-10-11. But (in the hope that this is not necessary) could you try the installation via conda?
