+++ .github/workflows/tests.yml
@@ -8,7 +8,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [3.5]
+        python-version: [3.7]
 
     steps:
       - name: Checkout github repo (+ download lfs dependencies)
