+++ .github/workflows/tests.yml
@@ -23,7 +23,7 @@ jobs:
           $CONDA/bin/conda install scipy
           $CONDA/bin/conda install -c anaconda cython
           $CONDA/bin/conda install -c conda-forge openmp
-          $CONDA/bin/conda install -c conda-forge openmpi=4.1.2
+          $CONDA/bin/conda install -c conda-forge openmpi
           $CONDA/bin/conda install matplotlib
           $CONDA/bin/conda install flake8
           $CONDA/bin/conda install -c conda-forge cythongsl
