### 2023-01-11 tibordome/cosmic_profiles#1 @benediktdiemer
@tibordome I've looked into this issue a bit more. I don't use conda, and tbh I don't wanna add a massive conda installation to my MacBook just for this. I checked that I can compile on my Mac with -fopenmp, and it works for gcc but not for clang; I believe both eventually call llvm, so the result should be the same. But where can I tell cosmic-profiles to compile using gcc rather than clang?
