### 2023-01-11 tibordome/cosmic_profiles#1 @tibordome
@benediktdiemer I have managed (in a docker macOS container) to install cosmic-profiles in conda (quite hefty in terms of disk space I agree) and also pip. The steps were
- brew install llvm
- fixing PATH, DYLD_LIBRARY_PATH etc. as outlined [here](https://blog.llvm.org/2015/05/openmp-support_22.html)
- downgrade (on the macOS Catalina that I was using) command line tools to the one for XCode 11.5 (solving [ld: unsupported tapi file type '!tapi-tbd' in YAML file](https://stackoverflow.com/questions/63592445/ld-unsupported-tapi-file-type-tapi-tbd-in-yaml-file) issue)
- `pip install cosmic-profiles` succeeds with clang

I think in your case, one could either set a symlink from clang to gcc (which I assume you have installed manually since the default /usr/bin/gcc points to clang) as described [here](https://github.com/tensorflow/ngraph-bridge/issues/446), or override the C/C++ compilers used during the build (same link, something like `CC=/path/to/gcc CXX=/path/to/gcc /path/to/pip install cosmic-profiles`).
