### 2023-01-11 tibordome/cosmic_profiles#1 @benediktdiemer
I seems that the issue is that /usr/bin/clang on MacOS always refers to the system version of clang, and this version somehow does not seem to play with libomp. I just updated all my compilers (llvm, clang15, libomp15), and I can compile openmp code with clang15 now. However, /usr/bin/clang still doesn't refer to the new clang, and even setting $CC does not help.

How can I force the pip installation to use a different compiler?
