### 2023-01-12 tibordome/cosmic_profiles#1 @benediktdiemer
I've changed my system such that clang points to a version that supports omp, as in, I can "clang -fopenmp test_omp.c -o test" compiles a test problem without errors.

Unfortunately, even when I have CC=clang, the cosmic-profiles installation still reverts specifically to /usr/bin/clang. This is the offending command:

      /usr/bin/clang -Wno-unused-result -Wsign-compare -Wunreachable-code -fno-common -dynamic -DNDEBUG -g -fwrapv -O3 -Wall -pipe -I/opt/local/libexec/openssl11/include -Os -isysroot/Library/Developer/CommandLineTools/SDKs/MacOSX10.15.sdk -I/private/tmp/pip-build-env-bx3fcp1w/overlay/lib/python3.7/site-packages/numpy/core/include -I. -I/opt/local/Library/Frameworks/Python.framework/Versions/3.7/include/python3.7m -c cosmic_profiles/cython_helpers/helper_class.c -o build/temp.macosx-10.15-x86_64-cpython-37/cosmic_profiles/cython_helpers/helper_class.o -fopenmp

Can you please test this on your machines? Is there something in the make files that forces this behavior?
