### 2023-01-13 tibordome/cosmic_profiles#1 @tibordome
I have tried again and can confirm that the ``CC=/path/to/clang/with/openmp/support`` flag works as expected. Specifically, the following throws an ``-fopenmp``-related error,

```
/Users/tibordome/opt/anaconda3/envs/jan13_pip/bin/pip install --force-reinstall --no-cache-dir cosmic-profiles
```

where I have added the ``--no-cache-dir`` flag since otherwise the pre-built binary would be fetched from the cache which would pass trivially. However, setting the correct ``CC`` flag using

```
CC=/usr/local/opt/llvm/bin/clang /Users/tibordome/opt/anaconda3/envs/jan13_pip/bin/pip install --force-reinstall --no-cache-dir cosmic-profiles
```

leads to a successful installation, where ``/usr/local/opt/llvm/bin/clang`` is the version of clang that provides OpenMP support (i.e. ``/usr/local/opt/llvm/bin/clang -fopenmp -lstdc++ test.cpp`` passes). As a final test,

```
CC=/usr/bin/clang /Users/tibordome/opt/anaconda3/envs/jan13_pip/bin/pip install --force-reinstall --no-cache-dir cosmic-profiles
```

fails again, as expected.
