### 2023-01-13 tibordome/cosmic_profiles#1 @benediktdiemer
Unfortunately, setting CC seems to work on some systems but not on others. Turns out the deeper reason is that the compiler is hard set in distutils.sysconfig:

https://shwina.github.io/custom-compiler-linker-extensions/

This can be verified in python:

>>> import distutils.sysconfig
>>> distutils.sysconfig.get_config_var("CC")
'/usr/bin/clang'

I will look into whether this can be changed.
