### 2023-01-13 tibordome/cosmic_profiles#1 @benediktdiemer
One addition: apparently the way to alter the distutils config is to add a ~/.pydistutils.cfg file in the home directory. Indeed, this seems to be processed, but I cannot just put a path to a compiler there, as in,

[build]
compiler=/opt/local/bin/clang

does not work. I'm not a specialist for distutils and I can't find good documentation. Help would be appreciated.
