### 2023-01-13 tibordome/cosmic_profiles#1 @tibordome
It's unfortunate that setting ``CC`` does not help on your machine. The ``~/.pydistutils.cfg`` solution would be a cleaner one I agree, though according to [this post](https://stackoverflow.com/questions/16077481/distutils-compiler-options-configuration), one should be able to override these configuration options by setting environment variables.

The best docs about distutils I have found are [here](https://pl.python.org/docs/inst/config-syntax.html) and [here](https://bouktin.blogspot.com/2012/04/configure-pydistutilscfg-python.html). According to [this post](https://github.com/pypa/setuptools/issues/3681), setting the path as you showed might not work, but instead one could try

```
[build]
compiler=clang15
```

i.e. with the respective clang command. I am worried though that this only affects invocations of ``python3 setup.py build`` etc., which during the building of the wheel (in a pip installation) might not be invoked after all. (I can try this at the weekend but at the moment the docker macOS container has eaten up too much disk space on my Linux machine which I will have to reorganise.)
