### 2023-01-13 tibordome/cosmic_profiles#1 @benediktdiemer
Indeed, on some people's system the env variables work and on some they don't; I am not sure why but it definitely doesn't work for me. I had found those same documentation pages, but they only say

--compiler (-c)    specify the compiler type

I've tried compiler=clang, which points to the correct executable; but I get

      running build_ext
      error: don't know how to compile C/C++ code on platform 'posix' with 'clang' compiler
      [end of output]

and the same for gcc. So I really don't understand what values are allowed.
