### 2023-01-17 tibordome/cosmic_profiles#1 @tibordome
I have checked and can confirm that I receive the same error messages when adding the compiler path or command name naively to the `[build]` section.

According to [this post](https://mail.python.org/pipermail/distutils-sig/2008-January/008608.html), the ``~./pydistutils.cfg`` file follows the nomenclature of ``python setup.py option=value``. I.e. having ``compiler=clang`` in the `[build]` section is equivalent to having ``python setup.py --compiler=clang`` during the building of the wheel when running ``pip install cosmic-profiles``, which will fail.

This is also illustrated [here](https://stackoverflow.com/questions/16737260/how-to-tell-distutils-to-use-gcc), which again suggests to set the ``CC`` environment variable. The ``CC`` variable should get checked by distutils, and I can confirm again on a clean macOS Catalina machine that it works as expected. If setting the ``CC`` environment variable does not work on your machine, I would suggest (as the [very same post](https://stackoverflow.com/questions/16737260/how-to-tell-distutils-to-use-gcc) does too) to set a symbolic link, which should definitely work as a last resort.
