### 2023-01-19 tibordome/cosmic_profiles#1 @benediktdiemer
Thanks for these suggestions. I'm currently moving and will start teaching a new class next week, so not totally sure when I'll get to playing with this.
