### 2023-02-07 #5008 @adonath
Just for the record: there is an unresolved installation issue, which prevents @benediktdiemer to install and run and review the software. @tibordome recommended a solution, which remains to be tested.

@benediktdiemer Please let me know in case you won't be able to proceed with the review any time soon (let's say if you need more > 2 weeks). I think we should make sure to have regular progress in the review process. In order to save time I could also try to look into the installation issue, or maybe find an expert from the JOSS side to help. Thanks!
