### 2023-02-08 #5008 @benediktdiemer
That's right. It's on my to-do list of course, but I've been unusually busy. The proposed solution (sym-linking my system compiler to something else) obviously has the potential to destroy my system if I don't do it right, so I need to take a little time deal with it. I do hope to get to it within the next two weeks.
