### 2023-03-04 #5008 @benediktdiemer
Alright, I've finally had time to take another stab at this -- sorry for the delay. I still cannot install cosmic-profiles on my macbook, even after disabling SIP and changing the /usr/bin/cc link manually. Even though pip uses gcc then (for some reason), it fails (I've checked that gcc can compile OpenMP). I've given up on getting cosmic-profiles to work on a Mac, and instead installed it on a different (unix) server.

I successfully ran some cosmic-profiles routines on the z = 0 snapshot of IllustrisTNG100-3-DM -- it worked more or less right out of the box! In particular, I managed to measure the shapes of the 10 largest halos and plot the shape parameters as well as visualizations, following the instructions in the docs. Unfortunately, I did not manage to run the examples for density profiles. Here, the interface is fairly complex. I got lost/stuck where the documentation says

"The first argument dens_profs_fit is an array of shape containing the density profiles defined at radii r_over_r200_fit, possibly obtained via getDensProfsDirectBinning() or getDensProfsDirectBinning(), with some non-reliable values removed."

The two suggested functions are the same, but they are not documented, and it is not clear how "non-reliable" values should be removed (or what they are). It would be much better to provide a fully worked set of commands that the user can test, and/or complete the documentation. By looking at the example script, I realized that the r_over_r200 array can just be generated with np.logspace or whatever.

Finally, I tried executing the "Gadget HDF5 Example" script. I could mostly run it, but it still fails on the density profiles part (meaning I still couldn't quite verify the density profiles functionality). I list some of the issues below. I have started confirming most items in the checklist above. Here is a list of issues that I'd like to see addressed before confirming the rest of the checks:

**Issues (code):**
- Complete documentation/example for density profiles
- DensShapeProfsHDF5() runs without error even when paths don't exist
- Visualization routines like plotShapeProfs() work, but the VIZ_DEST parameter needs documentation (I figured it out)
- Why does the user need to give the box size?
- In the example script, I had to change it from attempting to run over half the halos it finds to 20; otherwise the code crashes quickly (on a moderate machine, depending on the simulation)
- The script should automatically create the cat and viz dirs (otherwise errors occur)
- With that, the code still crashes on:

```
638	s: Starting fitDensProfs() with snap 099
best_fits = cprofiles.fitDensProfs(dens_profs[:,25:], ROverR200[25:], method = 'nfw', select = halos_select)
OSError: [Errno 12] Cannot allocate memory
(abbreviated)
```

NB this machine has 192 GB memory or so... not sure why this crashes. My suggestion is to try and design a minimal example script that can be run with reasonable resources (as long as the halos aren't too large).

**Issues (paper/docs):**
- In the Statement of Need, it would be good to more clearly distinguish between codes that extract density profiles from simulations and those that do fitting. I appreciate the reference to Colossus, but the text makes it sound as though it falls into the former category, when it only does the latter (for the record, I use SPARTA to extract profiles, but I'm not sure how relevant that is here).
- I wasn't sure whether there are any automated tests, as required by the checklist above?

**Suggestions for the future:**
- It is a little limiting that the code supports only Gadget/Arepo hdf5 and Subfind. For example, I have a lot of Gadget2 simulations (not in hdf5 format) and Rockstar catalogs. Obviously, adding more codes is a major effort, and I'm not sure to what extent generality is a goal here. But I think investing into a generalized interface such as cosmic-profiles makes sense only if it can be applied to a wide range of simulations.
