### 2023-03-07 tibordome/cosmic_profiles#1 @adonath
@tibordome To improve the pip install process you could maybe look into building and providing wheels for the package.
