### 2023-03-07 #5008 @adonath
Thanks a lot @benediktdiemer for the thorough review!

> I wasn't sure whether there are any automated tests, as required by the checklist above?

I think test are located here: https://github.com/tibordome/cosmic_profiles/tree/master/cosmic_profiles/tests
