### 2023-03-16 80f1287
Improved docs for estDensProfs() and fitDensProfs(); added doc for VIZ_DEST; raise ValueError if no HDF5 files found; raise ValueError if no subhalo found
