### 2023-03-16 1bbf22e
Create CAT_DEST and VIZ_DEST directories if not yet available
