+++ cosmic_profiles/dens_profs/dens_profs_tools.py
@@ -10,6 +10,9 @@ from functools import partial
 import os
 from cosmic_profiles.common.caching import np_cache_factory
 from scipy import optimize
+import inspect
+import subprocess
+currentdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
 from mpi4py import MPI
 comm = MPI.COMM_WORLD
 rank = comm.Get_rank()
@@ -150,6 +153,8 @@ def drawDensProfs(VIZ_DEST, SNAP, r200s, dens_profs_fit, ROverR200_fit, dens_pro
         best_fit, obj_nb = fitDensProf(ROverR200_fit, method, (prof_median_fit, r200, 0)) # Fit median
         best_fit_dict = getModelParsDict(best_fit, method)
         del r200
+        # Create VIZ_DEST if not available
+        subprocess.call(['mkdir', '-p', '{}'.format(VIZ_DEST)], cwd=os.path.join(currentdir))
         # Plotting
         plt.figure()
         plt.loglog(ROverR200_fit, prof_models[method](ROverR200_fit*np.average(r200s[np.arange(r200s.shape[0])]), best_fit_dict), 'o--', color = 'r', linewidth=2, markersize=4, label=r'${}$-profile fit'.format(model_name[method]))

[…]
