### 2023-03-16 13ecf29
Allow L_BOX = 0.0 and document meaning of non-periodic box; further improve doc on CAT_DEST, VIZ_DEST
