### 2023-03-29 8f00dcd
Turned 2D index catalogues, idx_cat and obj_cat, into 1D arrays; obj_numbers containing integer arrays instead of select; isValidSelection() updated
