+++ cosmic_profiles/common/python_routines.py
@@ -546,8 +546,8 @@ def calcCoM(xyz, masses):
 def getCatWithinFracR200(cat_in, obj_size_in, xyz, masses, L_BOX, CENTER, r200, frac_r200):
     """ Cleanse index catalogue ``cat_in`` of particles beyond R200 ``r200``
     
-    :param idx_cat_in: each row contains indices of particles belonging to an object
-    :type idx_cat_in: (N1, N3) integers
+    :param cat_in: contains indices of particles belonging to an object
+    :type cat_in: (N3) integers
     :param obj_size_in: indicates how many particles are in each object
     :type obj_size_in: (N1,) integers
     :param xyz: coordinates of particles of type 1 or type 4
@@ -566,40 +566,65 @@ def getCatWithinFracR200(cat_in, obj_size_in, xyz, masses, L_BOX, CENTER, r200, 
     :return cat_out, obj_size_out: updated cat_in and obj_size_out, with particles
         beyond R200 ``r200`` removed
     :rtype: list of length N1"""
-    cat_out = np.zeros((len(cat_in),np.max(obj_size_in)), dtype = np.int32)
-    obj_size_out = np.zeros((len(cat_in),), dtype = np.int32)
-    centers = np.zeros((len(cat_in),3), dtype = np.float32)
-    for idx in range(len(cat_in)): # Calculate centers of objects
-        xyz_ = respectPBCNoRef(xyz[cat_in[idx,:obj_size_in[idx]]], L_BOX)
+    cat_out = np.empty(0, dtype = np.int32)
+    obj_size_out = np.zeros((len(obj_size_in),), dtype = np.int32)
+    centers = np.zeros((len(obj_size_in),3), dtype = np.float32)
+    for idx in range(len(obj_size_in)): # Calculate centers of objects
+        xyz_ = respectPBCNoRef(xyz[cat_in[np.sum(obj_size_in[:idx]):np.sum(obj_size_in[:idx+1])]], L_BOX)
         if CENTER == 'mode':
-            centers.base[idx] = calcMode(xyz_, masses[cat_in[idx,:obj_size_in[idx]]], max((max(xyz_[:,0])-min(xyz_[:,0]), max(xyz_[:,1])-min(xyz_[:,1]), max(xyz_[:,2])-min(xyz_[:,2]))))
+            centers.base[idx] = calcMode(xyz_, masses[cat_in[np.sum(obj_size_in[:idx]):np.sum(obj_size_in[:idx+1])]], max((max(xyz_[:,0])-min(xyz_[:,0]), max(xyz_[:,1])-min(xyz_[:,1]), max(xyz_[:,2])-min(xyz_[:,2]))))
         else:
-            centers.base[idx] = calcCoM(xyz_, masses[cat_in[idx,:obj_size_in[idx]]])
+            centers.base[idx] = calcCoM(xyz_, masses[cat_in[np.sum(obj_size_in[:idx]):np.sum(obj_size_in[:idx+1])]])
     remnant = []
-    for idx, obj in enumerate(cat_in):
-        xyz_ = respectPBCNoRef(xyz[obj[:obj_size_in[idx]]])
+    for idx in range(len(obj_size_in)):
+        obj = cat_in[np.sum(obj_size_in[:idx]):np.sum(obj_size_in[:idx+1])]
+        xyz_ = respectPBCNoRef(xyz[obj])
         tree = cKDTree(xyz_, leafsize=2, balanced_tree = False)
         all_nn_idxs = tree.query_ball_point(centers[idx], r=r200[idx]*frac_r200, n_jobs=-1)
         if all_nn_idxs != []:
-            cat_out[idx] = list(np.array(obj[:obj_size_in[idx]])[all_nn_idxs])
[…]
