+++ example_scripts/apply_ahf.py
@@ -188,16 +188,18 @@ def AHFEx():
     #############################################################################################################
     
     ############## Run cosmic_profiles: define DensShapeProfs object ############################################
+    r_vir = np.float32([0.1, 0.1, 0.1, 0.1, 0.1])
     cprofiles = DensShapeProfs(dm_xyz, mass_array, h_indices, r_vir, SNAP, L_BOX, MIN_NUMBER_DM_PTCS, D_LOGSTART, D_LOGEND, D_BINS, IT_TOL, IT_WALL, IT_MIN, CENTER)
     if rank == 0:
-        h_idx_cat_len = len(cprofiles.getIdxCat()[0])
+        h_idx_cat_len = len(cprofiles.getIdxCat()[1])
     else:
         h_idx_cat_len = None
     h_idx_cat_len = comm.bcast(h_idx_cat_len, root = 0)
-    halos_select = [0, h_idx_cat_len//2]
-    
+    obj_numbers = [0, h_idx_cat_len//2]
+    print(r_vir)
+        
     ############## Create local halo shape catalogue ############################################################
-    cprofiles.dumpShapeCatLocal(CAT_DEST, select = halos_select, reduced = True, shell_based = True)
+    cprofiles.dumpShapeCatLocal(CAT_DEST, obj_numbers, reduced = True, shell_based = True)
     #############################################################################################################
     
     ############## Viz first halo ###############################################################################
