### 2023-03-29 5542829
No caching of RAM-heavy functions, reloading from CAT_DEST instead
