+++ cosmic_profiles/shape_profs/shape_profs_algos.pyx
@@ -514,7 +514,6 @@ cdef float[:] runShellVDispAlgo(float[:] morph_info, float[:,:] xyz, float[:,:] 
     return morph_info
 
 @cython.embedsignature(True)
-@cython.binding(True)
 def calcMorphLocal(float[:,:] xyz, float[:] masses, float[:] r200, int[:] idx_cat, int[:] obj_size, float L_BOX, int D_LOGSTART, int D_LOGEND, int D_BINS, int IT_TOL, int IT_WALL, int IT_MIN, str CENTER, bint reduced, bint shell_based):
     """ Calculates the local shape catalogue
     
@@ -557,86 +556,80 @@ def calcMorphLocal(float[:,:] xyz, float[:] masses, float[:] r200, int[:] idx_ca
     :return: d, q, s, eigframe, centers, masses, l_succeed: list of object indices for which morphology could be determined at R200 (length: N3)
     :rtype: (N3, ``D_BINS`` + 1) floats (for d, q, s, eigframe (x3)), (N3, 3) floats (for centers), (N3,) floats (for masses), N3-list of ints for l_succeed
     """
-    def inner(float[:,:] xyz, float[:] masses, float[:] r200, int[:] idx_cat, int[:] obj_size, float L_BOX, int D_LOGSTART, int D_LOGEND, int D_BINS, int IT_TOL, int IT_WALL, int IT_MIN, str CENTER, bint reduced, bint shell_based):
-        # Transform cat to int[:,:]
-        cdef int nb_objs = obj_size.shape[0]
-        cdef int[:] offsets = np.int32(np.hstack((np.array([0]), np.cumsum(obj_size.base))))
-        cdef int p
-        cdef int largest_size = np.max(obj_size.base)
-        cdef float[:] m = np.zeros((nb_objs,), dtype = np.float32)
-        cdef float[:,:] d = np.zeros((nb_objs, D_BINS+1), dtype = np.float32)
-        cdef float[:,:] q = np.zeros((nb_objs, D_BINS+1), dtype = np.float32)
-        cdef float[:,:] s = np.zeros((nb_objs, D_BINS+1), dtype = np.float32)
-        cdef float[:,:] major_x = np.zeros((nb_objs, D_BINS+1), dtype = np.float32)
-        cdef float[:,:] major_y = np.zeros((nb_objs, D_BINS+1), dtype = np.float32)
-        cdef float[:,:] major_z = np.zeros((nb_objs, D_BINS+1), dtype = np.float32)
-        cdef float[:,:] inter_x = np.zeros((nb_objs, D_BINS+1), dtype = np.float32)
-        cdef float[:,:] inter_y = np.zeros((nb_objs, D_BINS+1), dtype = np.float32)
-        cdef float[:,:] inter_z = np.zeros((nb_objs, D_BINS+1), dtype = np.float32)
-        cdef float[:,:] minor_x = np.zeros((nb_objs, D_BINS+1), dtype = np.float32)
-        cdef float[:,:] minor_y = np.zeros((nb_objs, D_BINS+1), dtype = np.float32)
-        cdef float[:,:] minor_z = np.zeros((nb_objs, D_BINS+1), dtype = np.float32)
-        cdef float[:,:,:] morph_info = np.zeros((openmp.omp_get_max_threads(), 12, D_BINS+1), dtype = np.float32)
-        cdef float[:,:,:] xyz_obj = np.zeros((openmp.omp_get_max_threads(), largest_size,3), dtype = np.float32)
-        cdef float[:,:,:] xyz_princ = np.zeros((openmp.omp_get_max_threads(), largest_size,3), dtype = np.float32)
[…]

[…]
