### 2023-04-11 ef30852
Restructured so that GADGET HDF5 has simple interface, removed redundancies. Since multiple inheritance would lead to instance lay-out conflict, copy HDF5 methods from DensProfsHDF5 to DensShapeProfsHDF5, i.e. small redundancy remains
