+++ cosmic_profiles/dens_profs/dens_profs_classes.pyx
@@ -7,34 +7,36 @@ from cosmic_profiles.common.cosmic_base_class cimport CosmicBase
 from cosmic_profiles.common import config
 from cosmic_profiles.common.python_routines import print_status, isValidSelection, getSubSetIdxCat
 from cosmic_profiles.dens_profs.dens_profs_tools import drawDensProfs
-from cosmic_profiles.gadget_hdf5.get_hdf5 import getHDF5SHData, getHDF5ObjData
+from cosmic_profiles.gadget_hdf5.get_hdf5 import getHDF5SHData, getHDF5ObjData, getPartType
 from cosmic_profiles.gadget_hdf5.gen_catalogues import calcObjCat
 import time
 from mpi4py import MPI
 comm = MPI.COMM_WORLD
 rank = comm.Get_rank()
 size = comm.Get_size()
-
-cdef class DensProfs(CosmicBase):
+        
+cdef class DensProfsBase(CosmicBase):
     """ Class for density profile calculations
     
     Its public methods are ``getR200()``, ``getIdxCat()``,
     ``getXYZMasses()``, ``getMassesCenters()``, ``_getMassesCenters()``, ``estDensProfs()``, 
     ``fitDensProfs()``, ``estConcentrations()``, ``plotDensProfs()``."""
     
-    def __init__(self, float[:,:] xyz, float[:] masses, idx_cat, float[:] r200, str SNAP, float L_BOX, int MIN_NUMBER_PTCS, str CENTER, str VIZ_DEST, str CAT_DEST):
+    def __init__(self, float[:,:] xyz, float[:] masses, int[:] idx_cat, float[:] r200, int[:] obj_size, str SNAP, float L_BOX, int MIN_NUMBER_PTCS, str CENTER, str VIZ_DEST, str CAT_DEST, str SUFFIX):
         """
-        :param xyz: positions of all simulation particles in config.InUnitLength_in_cm
+        :param xyz: positions of all simulation particles in Mpc/h
         :type xyz: (N2,3) floats, N2 >> N1
-        :param masses: masses of all simulation particles in config.config.InUnitMass_in_g
+        :param masses: masses of all simulation particles in 10^10 M_sun/h
         :type masses: (N2,) floats
         :param idx_cat: each entry of the list is a list containing indices of particles belonging to an object
         :type idx_cat: list of length N1
-        :param r200: R_200 radii of the parent halos in config.InUnitLength_in_cm
+        :param r200: R_200 radii of the parent halos in Mpc/h
         :type r200: (N1,) floats
+        :param obj_size: indicates how many particles are in each object
+        :type obj_size: (N1,) integers
         :param SNAP: snapshot identifier, e.g. '024'
         :type SNAP: string
-        :param L_BOX: simulation box side length in config.InUnitLength_in_cm
+        :param L_BOX: simulation box side length in Mpc/h
         :type L_BOX: float
         :param MIN_NUMBER_PTCS: minimum number of particles for object to qualify for morphology calculation
         :type MIN_NUMBER_PTCS: int
@@ -44,31 +46,16 @@ cdef class DensProfs(CosmicBase):
         :param VIZ_DEST: visualization folder
         :type VIZ_DEST: string
         :param CAT_DEST: catalogue destination
-        :type CAT_DEST: string"""
-        super().__init__(SNAP, L_BOX, MIN_NUMBER_PTCS, CENTER, VIZ_DEST, CAT_DEST)
+        :type CAT_DEST: string
+        :param SUFFIX: either '_dm_' or '_gx_' or '_' (latter for CosmicProfsDirect)
+        :type SUFFIX: string"""
[…]
