+++ cosmic_profiles/shape_profs/shape_profs_classes.pyx
@@ -3,10 +3,10 @@
 # -*- coding: utf-8 -*-
 
 import numpy as np
+from numpy.random import default_rng
 import matplotlib.pyplot as plt
 from mpl_toolkits.mplot3d import Axes3D
 from matplotlib import pyplot
-from numpy.random import default_rng
 from matplotlib.font_manager import FontProperties
 import matplotlib
 matplotlib.rcParams.update({'font.size': 13})
@@ -15,18 +15,19 @@ import inspect
 import subprocess
 import os
 currentdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
-from cosmic_profiles.dens_profs.dens_profs_classes cimport DensProfs, DensProfsHDF5
+from cosmic_profiles.dens_profs.dens_profs_classes cimport DensProfsBase, DensProfsHDF5
 from cosmic_profiles.common.python_routines import print_status, set_axes_equal, fibonacci_ellipsoid, respectPBCNoRef, isValidSelection, getSubSetIdxCat
 from cosmic_profiles.common import config
 from cosmic_profiles.shape_profs.shape_profs_tools import getGlobalEpsHist, getLocalEpsHist
-from cosmic_profiles.gadget_hdf5.get_hdf5 import getHDF5SHData
+from cosmic_profiles.gadget_hdf5.get_hdf5 import getHDF5SHData, getHDF5ObjData, getPartType
+from cosmic_profiles.gadget_hdf5.gen_catalogues import calcObjCat
 import time
 from mpi4py import MPI
 comm = MPI.COMM_WORLD
 rank = comm.Get_rank()
 size = comm.Get_size()
       
-cdef class DensShapeProfs(DensProfs):
+cdef class DensShapeProfsBase(DensProfsBase):
     """ Class for density profile and shape profile calculations
     
     Its public methods are ``estDensProfs()``, ``getShapeCatLocal()``, ``getShapeCatGlobal()``, 
@@ -41,7 +42,7 @@ cdef class DensShapeProfs(DensProfs):
     cdef int IT_WALL
     cdef int IT_MIN
     
-    def __init__(self, float[:,:] xyz, float[:] masses, idx_cat, float[:] r200, str SNAP, float L_BOX, int MIN_NUMBER_PTCS, int D_LOGSTART, int D_LOGEND, int D_BINS, float IT_TOL, int IT_WALL, int IT_MIN, str CENTER, str VIZ_DEST, str CAT_DEST):
+    def __init__(self, float[:,:] xyz, float[:] masses, idx_cat, float[:] r200, int[:] obj_size, str SNAP, float L_BOX, int MIN_NUMBER_PTCS, int D_LOGSTART, int D_LOGEND, int D_BINS, float IT_TOL, int IT_WALL, int IT_MIN, str CENTER, str VIZ_DEST, str CAT_DEST, str SUFFIX):
         """
         :param xyz: positions of all simulation particles in config.InUnitLength_in_cm
         :type xyz: (N2,3) floats, N2 >> N1
@@ -51,6 +52,8 @@ cdef class DensShapeProfs(DensProfs):
         :type idx_cat: list of length N1
         :param r200: R_200 radii of the parent halos in config.InUnitLength_in_cm
         :type r200: (N1,) floats
+        :param obj_size: indicates how many particles are in each object
+        :type obj_size: (N1,) integers
         :param SNAP: snapshot identifier, e.g. '024'
         :type SNAP: string
         :param L_BOX: simulation box side length in config.InUnitLength_in_cm
@@ -67,7 +70,7 @@ cdef class DensShapeProfs(DensProfs):
             for iteration to stop
         :type IT_TOL: float
         :param IT_WALL: maximum permissible number of iterations
[…]
