### 2023-04-12 d96a9dc
Added Gadget I and II compatibility, Gadget type is recognized automatically; removed MASS_UNIT redundancies; L_BOX is now read off rather than provided by the user; improved docs
