+++ cosmic_profiles/for_docs/gadget/readgadget.py
@@ -1,0 +1,192 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Wed Apr 12 09:58:23 2023
+
+@author: tibor, franciscovillaescusa 
+"""
+
+# This library is designed to read Gadget format I and II formats, adapted from pylians, https://github.com/franciscovillaescusa/Pylians
+import numpy as np
+from gadget import readsnap
+import os,h5py
+from common import config
+
+# Find snapshot name and format
+def fname_format(snapshot):
+    if os.path.exists(snapshot):
+        if snapshot[-4:]=='hdf5':
+            filename, fformat = snapshot, 'hdf5'
+        else:                      
+            filename, fformat = snapshot, 'binary'
+    elif os.path.exists(snapshot+'.0'):
+        filename, fformat = snapshot+'.0', 'binary'
+    elif os.path.exists(snapshot+'.hdf5'):
+        filename, fformat = snapshot+'.hdf5', 'hdf5'
+    elif os.path.exists(snapshot+'.0.hdf5'):
+        filename, fformat = snapshot+'.0.hdf5', 'hdf5'
+    else:  
+        raise Exception('File not found!')
+    return filename,fformat
+
+
+# This class reads the header of the gadget file
+class header:
+    def __init__(self, snapshot):
+
+        filename, fformat = fname_format(snapshot)
+
+        if fformat=='hdf5':
+            f             = h5py.File(filename, 'r')
+            self.time     = f['Header'].attrs[u'Time']
+            self.redshift = f['Header'].attrs[u'Redshift']
+            self.boxsize  = f['Header'].attrs[u'BoxSize']
+            self.filenum  = f['Header'].attrs[u'NumFilesPerSnapshot']
+            self.omega_m  = f['Header'].attrs[u'Omega0']
+            self.omega_l  = f['Header'].attrs[u'OmegaLambda']
+            self.hubble   = f['Header'].attrs[u'HubbleParam']
+            self.massarr  = f['Header'].attrs[u'MassTable']
+            self.npart    = f['Header'].attrs[u'NumPart_ThisFile']
+            self.nall     = f['Header'].attrs[u'NumPart_Total']
+            self.cooling  = f['Header'].attrs[u'Flag_Cooling']
+            self.format   = 'hdf5'
+            f.close()
+
+        else:        
+            head = readsnap.snapshot_header(filename)
+            self.time     = head.time
+            self.redshift = head.redshift
+            self.boxsize  = head.boxsize
+            self.filenum  = head.filenum
+            self.omega_m  = head.omega_m
+            self.omega_l  = head.omega_l
+            self.hubble   = head.hubble
+            self.massarr  = head.massarr
+            self.npart    = head.npart
+            self.nall     = head.nall
+            self.cooling  = head.cooling
+            self.format   = head.format
+
+        # km/s/(Mpc/h)
+        self.Hubble = 100.0*np.sqrt(self.omega_m*(1.0+self.redshift)**3+self.omega_l)
+
+
+# This function reads a block of an individual file of a gadget snapshot
+def read_field(snapshot, block, ptype):
+
+    filename, fformat = fname_format(snapshot)
+    head              = header(filename)
+    Masses            = head.massarr      # In units of config.InUnitMass_in_g               
+    Npart             = head.npart        # Number of particles in the subfile
+
+    l_target_over_curr = 3.085678e24/config.InUnitLength_in_cm
+    m_target_over_curr = 1.989e43/config.InUnitMass_in_g
+    v_target_over_curr = 1e5/config.InUnitVelocity_in_cm_per_s
+
+    if block=="POS ":  
+        suffix = "Coordinates"
+        unit_fac = l_target_over_curr
+    elif block=="MASS":  
[…]
