+++ cosmic_profiles/dens_profs/dens_profs_classes.pyx
@@ -7,8 +7,9 @@ from cosmic_profiles.common.cosmic_base_class cimport CosmicBase
 from cosmic_profiles.common import config
 from cosmic_profiles.common.python_routines import print_status, isValidSelection, getSubSetIdxCat
 from cosmic_profiles.dens_profs.dens_profs_tools import drawDensProfs
-from cosmic_profiles.gadget_hdf5.get_hdf5 import getHDF5SHData, getHDF5ObjData, getPartType
-from cosmic_profiles.gadget_hdf5.gen_catalogues import calcObjCat
+from cosmic_profiles.gadget.read_fof import getFoFSHData, getPartType
+from cosmic_profiles.gadget.gen_catalogues import calcObjCat
+from cosmic_profiles.gadget import readgadget
 import time
 from mpi4py import MPI
 comm = MPI.COMM_WORLD
@@ -20,7 +21,7 @@ cdef class DensProfsBase(CosmicBase):
     
     Its public methods are ``getR200()``, ``getIdxCat()``,
     ``getXYZMasses()``, ``getMassesCenters()``, ``_getMassesCenters()``, ``estDensProfs()``, 
-    ``fitDensProfs()``, ``estConcentrations()``, ``plotDensProfs()``."""
+    ``fitDensProfs()``, ``estConcentrations()``, ``plotDensProfs()``, ``getObjInfo()``."""
     
     def __init__(self, float[:,:] xyz, float[:] masses, int[:] idx_cat, float[:] r200, int[:] obj_size, str SNAP, float L_BOX, int MIN_NUMBER_PTCS, str CENTER, str VIZ_DEST, str CAT_DEST, str SUFFIX):
         """
@@ -28,8 +29,8 @@ cdef class DensProfsBase(CosmicBase):
         :type xyz: (N2,3) floats, N2 >> N1
         :param masses: masses of all simulation particles in 10^10 M_sun/h
         :type masses: (N2,) floats
-        :param idx_cat: each entry of the list is a list containing indices of particles belonging to an object
-        :type idx_cat: list of length N1
+        :param idx_cat: contains indices of particles belonging to an object
+        :type idx_cat: (N3,) integers
         :param r200: R_200 radii of the parent halos in Mpc/h
         :type r200: (N1,) floats
         :param obj_size: indicates how many particles are in each object
@@ -70,9 +71,9 @@ cdef class DensProfsBase(CosmicBase):
     def getIdxCat(self): # Public Method
         """ Fetch catalogue
         
-        :return idx_cat: each row contains indices of particles belonging to an object,
+        :return idx_cat: contains indices of particles belonging to an object,
             obj_size: number of particles in each object
-        :rtype: (N1, N3) integers and (N1,) integers"""
+        :rtype: (N3) integers and (N1,) integers"""
         if rank == 0:
             return self.idx_cat.base, self.obj_size.base
         else:
@@ -86,8 +87,8 @@ cdef class DensProfsBase(CosmicBase):
         :rtype: (N2,3) floats, (N2,) floats"""
         if rank == 0:
             l_curr_over_target = 3.085678e24/config.OutUnitLength_in_cm
-            m_curr_over_target = 1.989e33/config.OutUnitMass_in_g
-            return self.xyz.base*l_curr_over_target, self.masses.base*self.MASS_UNIT*m_curr_over_target
+            m_curr_over_target = 1.989e43/config.OutUnitMass_in_g
+            return self.xyz.base*l_curr_over_target, self.masses.base*m_curr_over_target
         else:
             return None, None
[…]
