+++ cosmic_profiles/for_docs/dens_profs/dens_profs_tools.py
@@ -96,7 +96,7 @@ def drawDensProfs(VIZ_DEST, SNAP, r200s, dens_profs_fit, ROverR200_fit, dens_pro
     :type start_time: float
     :param MASS_UNIT: conversion factor from previous mass unit to M_sun/h, usually 10^10
     :type MASS_UNIT: float
-    :param suffix: either '_dm_' or '_gx_' or '' (latter for CosmicProfsDirect)
+    :param suffix: either '_dm_' or '_gx_' or '_' (latter for CosmicProfsDirect)
     :type suffix: string"""
     return
 
@@ -136,5 +136,4 @@ def fitDensProfHelper(dens_profs, ROverR200, r200s, method):
     :type method: string, either `einasto`, `alpha_beta_gamma`, `hernquist`, `nfw`
     :return best_fits: best-fit results
     :rtype: (N,n) floats"""
-    
     return
