+++ docs/run_doc_for_commit.sh
@@ -1,0 +1,4 @@
+cd ../cosmic_profiles/for_docs
+python3 setup_compile.py build_ext --inplace
+cd ../../docs
+sphinx-build -b html source build

[…]
