### 2023-04-15 163abdf
Improved density and shape tests using parametrized tests (in pytest); improved docs / paper
