+++ paper/paper.md
@@ -34,32 +34,33 @@ without a proper treatment thereof.
 The analysis of simulation outputs that inform observational searches requires
 reliable and fast numerical tools. `CosmicProfiles` is a Python package with a
 substantial Cython component to enable quick and easy calculation of global and
-local density as well as shape properties of finitely resolved objects. The shape
-profiles of the objects can in turn be used to improve the fidelity of density
-profiles by considering ellipsoidal shells instead of the popular choice of
-spherical shells. While `Colossus` [@Diemer_2018] provides some density profile
-fitting functionalities, `CosmicProfiles` in addition offers spherical shell-based
-and ellipsoidal shell-based density profile estimation techniques to find such
-profiles in the first place.
+local density as well as shape properties of finitely resolved objects. 
+Existing codes to extract density profiles include `SPARTA` [@Diemer_2022] 
+while density profile fitting functionalities are provided by 
+e.g. `Colossus` [@Diemer_2018]. The strength of `CosmicProfiles` lies in shape profiling.
+The shape profiles of the objects can in turn be used to improve the fidelity of density
+profiles by considering ellipsoidal shells (defined via shape profiles)
+instead of the popular choice of spherical shells as the basis for the 
+density profile extraction.
 
-The objects under consideration can either be halos or galaxies from a cosmological
-simulation but also point clouds from any other scientific research field. The API
-for `CosmicProfiles` was designed to provide a class-based and user-friendly interface
-to Cython-optimized implementations of common operations such as the estimation of
-density profiles and subsequent fitting thereof to the user's preferred density
-profile model. Interfaces to `pynbody` [@Pontzen_2013] and `nbodykit` [@Hand_2018]
-have been rendered very simple with detailed example scripts, such that e.g.
-halos that have been identified via `pynbody` can be fed to `CosmicProfiles`
-for radial profiling.
+The objects under consideration can either be dark matter / gas halos or galaxies
+(stellar particles) from a cosmological simulation but also point clouds 
+from any other scientific research field. The API for `CosmicProfiles` was designed 
+to provide a class-based and user-friendly interface to Cython-optimized implementations 
+of common operations such as the estimation of density profiles and subsequent 
+fitting thereof to the user's preferred density profile model. Interfaces to 
+`pynbody` [@Pontzen_2013] and `nbodykit` [@Hand_2018] have been rendered very 
+simple with detailed example scripts, such that e.g. halos that have been identified 
+via `pynbody` can be fed to `CosmicProfiles` for radial profiling.
 
 If no halos are available to the user, `CosmicProfiles` offers a versatile mock
 halo generator that will sample particles from a target density profile model and
 ellipsoidal shape distribution, all provided by the user. `CosmicProfiles`
 was designed to be used by both astronomical researchers and researchers of other
-fields that come across ellipsoidal point clouds. It has already been used in
-a scientific publication [@Dome_2022]. The combination of speed and design
+fields that come across spherical or ellipsoidal point clouds. It has already been used in
+a scientific publication [@Dome_2023]. The combination of speed and design
 will hopefully ease the post-processing of simulation snapshots such as Gadget-style
-HDF5 files [@Springel_2010] and contribute to exciting scientific explorations
+I, II or HDF5 files [@Springel_2010] and contribute to exciting scientific explorations
 by students and experts alike.
 
 # Acknowledgements
