+++ cosmic_profiles/tests/test_densities.py
@@ -7,6 +7,8 @@ Created on Fri Jul  1 16:21:27 2022
 
 import numpy as np
 import os
+import itertools
+import pytest
 import subprocess
 import sys
 import inspect
@@ -22,7 +24,8 @@ comm = MPI.COMM_WORLD
 rank = comm.Get_rank()
 size = comm.Get_size()
 
-def test_densities():
+@pytest.mark.parametrize('method, direct_binning', [p for p in itertools.product(*[['einasto', 'nfw', 'hernquist', 'alpha_beta_gamma'], [False, True]])])
+def test_densities(method, direct_binning):
     #################################### Parameters #################################################
     updateInUnitSystem(in_unit_length_in_cm = 3.085678e24, in_unit_mass_in_g = 1.989e33, in_unit_velocity_in_cm_per_s = 1e5)
     updateOutUnitSystem(out_unit_length_in_cm = 3.085678e24, out_unit_mass_in_g = 1.989e33, out_unit_velocity_in_cm_per_s = 1e5)
@@ -33,15 +36,16 @@ def test_densities():
     r_over_rvir = np.logspace(-2,0,50)
     VIZ_DEST = "./cosmic_profiles/tests/viz"
     CAT_DEST = "./cosmic_profiles/tests/cat"
-    method = 'einasto'
     nb_model_pars = {'einasto': 3, 'nfw': 2, 'hernquist': 2, 'alpha_beta_gamma': 5}
     N = 10 # Number of halos used for test
     
     #################################### Generate N mock halos ######################################
     r_s = 0.5 # Units are Mpc/h
     alpha = 0.18
+    beta = 3.0
+    gamma = 1.0
     
-    model_pars = {'alpha': alpha, 'r_s': r_s}
+    model_pars = {'einasto': {'alpha': alpha, 'r_s': r_s}, 'nfw': {'r_s': r_s}, 'hernquist': {'r_s': r_s}, 'alpha_beta_gamma': {'alpha': alpha, 'beta': beta, 'gamma': gamma, 'r_s': r_s}}
     dm_x = np.empty(0, dtype = np.float32)
     dm_y = np.empty(0, dtype = np.float32)
     dm_z = np.empty(0, dtype = np.float32)
@@ -54,7 +58,7 @@ def test_densities():
         a = np.logspace(-1.5,0.2,100)*r_vir[-1] # Units are Mpc/h
         b = a*0.6 # Units are Mpc/h
         c = a*0.2 # Units are Mpc/h
-        halo_x, halo_y, halo_z, mass_dm, rho_s = genHalo(tot_mass, halo_res, model_pars, 'einasto', a, b, c)
+        halo_x, halo_y, halo_z, mass_dm, rho_s = genHalo(tot_mass, halo_res, model_pars[method], method, a, b, c)
         print("Number of particles in the halo is {}.".format(halo_x.shape[0]))
         halo_x += np.random.uniform(0,L_BOX/2,1)[0] # Move mock halo into the middle of the simulation box
         halo_y += np.random.uniform(0,L_BOX/2,1)[0]
@@ -74,22 +78,19 @@ def test_densities():
     
     ############################## Estimate Density Profiles #########################################
     obj_numbers = np.arange(5)
-    dens_profs_db = cprofiles.estDensProfs(r_over_rvir, obj_numbers = obj_numbers, direct_binning = True) # dens_profs_db is in M_sun*h^2/Mpc^3
-    dens_profs_kb = cprofiles.estDensProfs(r_over_rvir, obj_numbers = obj_numbers, direct_binning = False) # These estimates will be kernel-based
+    dens_profs = cprofiles.estDensProfs(r_over_rvir, obj_numbers = obj_numbers, direct_binning = direct_binning) # dens_profs is in M_sun*h^2/Mpc^3
[…]

[…]
