+++ README.rst
@@ -18,7 +18,8 @@ This repository provides shape and density profile analysis tools for cosmologic
 - supports
 
   - 'direct' datasets (i.e. index catalogue provided by user) and
-  - Gadget-style HDF5 snapshot files
+  - FoF / SUBFIND halo catalogues
+  - Gadget-style I, II and HDF5 snapshot files
   
     - all functionalities available for dark matter halos, gas particle halos and star particle halos
     - in addition, allows for velocity dispersion tensor eigenaxes determination
