+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 1.3.0
+current_version = 1.3.1
 
 [build_sphinx]
 source-dir = docs/source
