### 2023-04-15 #5008 @tibordome
I should have addressed all comments / suggestions in version 1.3.1.

  - updated docs
    - density profile documentation
    - VIZ_DEST documentation
    - distinguish between density profile extraction and fitting algorithms
  - create cat and viz directories automatically
  - box size read off from Gadget I, II or HDF5 file rather than having to be provided
  - less RAM-heavy
    - it could be traced back to inefficient storing of particle index catalogue in a 2D array which had a lot of redundant zeros since not all objects (halos) have the same size
    - now I have a 1D index catalogue, and the object sizes provide the relevant offsets in the array
  - automated tests improved using parametrized tests (in pytest)
  - made compatible with Gadget-style I, II and HDF5 files
    - The snapshot reading functionalities are in parts based on Francisco Villaescusa-Navarro's open-source Pylians package, which I have acknowledged in the paper docs
  - I started working on Rockstar compatibility, but the .list files have no particle indices (of the halos) as far as I can tell, so I wasn’t sure how to construct the particle index catalogue. I know there are particle files for AHF, for instance. In any case, it would take more effort to make CosmicProfiles compatible with Rockstar, AHF etc.
