### 2023-04-15 tibordome/cosmic_profiles#1 @tibordome
@adonath A built wheel (.whl file) is provided for each version of the package on the pypi index, in addition to the source code (tar.gz file)
