### 2023-04-26 #5008 @phil-mansfield
Since Benedikt has already tested out running this on the pre-packaged particle formats, I tried my hand at getting this to run on some zoom-in data via `DensShapeProfs`. I haven't managed to get this to work yet, but here are my notes so far. Of the things I mention below, only the last is something that's stopping me from continuing.

Installing
---------

I was able to install the software without much difficulty on one of Stanford's clusters after some version updates. It seems that there's a version requirement where a sub-package called scipy._lib needs to exist which isn't enforced by pip. I don't know what version of scipy this corresponds to, but the most recent one is fine. I haven't been successful in getting it to work on my mac laptop yet, but haven't tried too hard to debug that yet.

When importing cosmic profiles on our cluster, if I do it within a scheduled SLURM job, I get the error message
```
*** An error occurred in MPI_Init_thread
*** on a NULL communicator
*** MPI_ERRORS_ARE_FATAL (processes in this communicator will now abort,
***    and potentially your MPI job)
[rome0182:88971] Local abort before MPI_INIT completed completed successfully, but am not able to aggregate error messages, and not able to guarantee that all other processes were killed!
```

I am able to successfully import the code from a login node (although it takes some time, close to a minute the first time I import it during a session and about 6 seconds after that).

Supported Data Structures page
--------------------------------

The "SNAP" variable is defined twice in the second code block

The comment on L_BOX  in the code blocks says it's in cMpc/h. When I read this I wasn't sure if it was in units of length_in_cm, of if all box sizes needed to be in cMpc/h, although this is specified later.

Not all simulations are run in periodic boxes. Also, in zoom-in simulations it's common to recenter the simulation on the main halo (i.e. so the halo is at (0, 0, 0) at z=0). I was testing this on some zoom-in data and needed to manually add in some arbitrarily large offset. This is not a big deal, though.

For `updateInUnitSystem` and `updateOutUnitSystem`:
- The "in_unit_" and "out_unit_" prefixes on the named arguments contain information which is already in the function names, so they could be renamed `length_in_cm`/etc. to reduce the amount of typing.
- It might be a nice convenience feature to allow the user to pass strings corresponding to commonly used units (eg. length_in_cm="kpc" or mass_in_g="Msun") or something similar.
- In my opinion, it would be better to make these methods on the two DensShapeProfs classes rather than top-level functions which set global state. (Or classes which can be initialized and passed as arguments to the class intializers). If I have two different DensShapeProf classes that have different units (e.g. two different simulations, or two from the same simulation at different scale factors), they'll need different unit conversions. If I'm analyzing them at the same time, would I need to re-call both functions each time I'm switching from one set of shape profiles to the other? It seems like it would be less likely to result in user error if these were directly tied to classes.
- More generally, it might just be cleaner to have a single unit convention used by the entire library for outputs. Alternately, I think assuming that the user would want to use the same units for inputs and outputs is also pretty safe.

The class initializers have a very large number of arguments. I think it would be best to decrease this: it would make it less overwhelming to new users and less likely for people to put in errors. Might be best to make D_LOGSTART, D_LOGEND, D_BINS, IT_TOL, IT_WALL, IT_MIN, CENTER, and MIN_NUMBER_PTCS named arguments. Additionally, it seems like the SNAP, GROUP_DEST, and VIZ_DEST arguments might only be needed if the user calls certain functions to dump the binary profiles/plots. I'd recommend attaching them to the I/O functions that need them.

The shape of xyz isn't specified. (i.e. is it (3,N) or (N,3))

When I try to initialize DensShapeProfs with the example code in the documentation, I get the error
```
TypeError: __init__() takes exactly 16 positional arguments (14 given)
```
I think it's missing VIZ_DEST and CAT_DEST based on the docstring for the initializer.

My data was originally stored as doubles and this led to an error when passed to DensShapeProfs.
```
ValueError: Buffer dtype mismatch, expected 'float' but got 'double'
```
The required type isn't specified in the documentation, and it might be best to call, e.g. np.asarray() to do the type conversion internally.

The docstring for the class's intializer says that D_LOGSTART and D_LOGEND must be integers, but I don't think this is mentioned in the online documentation. (Additionally, I think a user could pretty reasonably want the logarithmic bins to start/end at something other than a power of ten)

With these issues fixed, I get the following error when I try to call the initializer.
```
    cprofiles = DensShapeProfs(xyz, mass_array, idx_cat, r_vir, SNAP, L_BOX, MIN_NUMBER_PTCS, D_LOGSTART, D_LOGEND, D_BINS, IT_TOL, IT_WALL, IT_MIN, CENTER, VIZ_DEST, CAT_DEST)
  File "cosmic_profiles/shape_profs/shape_profs_classes.pyx", line 459, in cosmic_profiles.shape_profs.shape_profs_classes.DensShapeProfs.__init__
  File "stringsource", line 660, in View.MemoryView.memoryview_cwrapper
  File "stringsource", line 350, in View.MemoryView.memoryview.__cinit__
```
I've checked the `.flag` field for the three numpy arryas I'm passing to DensShapeProfs (xyz, mass_array, and idx_cat) and each of them has WRITEABLE = True. Do you know what's happening here?
