### 2023-04-27 #5008 @tibordome
Many thanks for the suggestions and comments, I'll start addressing them.

I think the

``Local abort before MPI_INIT completed completed successfully``

error message might be an MPI issue on the cluster itself, based on [this thread](https://github.com/3dem/relion/issues/118). I usually do ``conda install -c conda-forge mpi4py`` before the actual install (either pip or conda) on any machine, perhaps I should add that to the documentation.

In a completely empty environment on a Linux machine, I can do ``pip install --user [cosmic-profiles``](url) with no issues regarding scipy. On which platform does this occur?

I assume the actual error message is

``ValueError: buffer source array is read-only``

I've read somewhere that this might occur if the arrays are too large, but I increased the arrays to 1 billion \times 3 (for dm_xyz), and still no such error. I've tried on a cluster here in Cambridge and I couldn't reproduce it there either. Could you try this example script that you will find attached?
[ex_script.zip](https://github.com/openjournals/joss-reviews/files/11342361/ex_script.zip)
