### 2023-04-27 #5008 @phil-mansfield
I will try these out later today. Also, I forgot to mention this earlier, but all the comments about function style and that type of thing are pure suggestions. I trust your judgement on whether or not you want to follow them and it's okay if you end up deciding to ignore them. I'll demarcate those more clearly going forward.
