### 2023-04-27 #5008 @phil-mansfield
Another quick suggestion. This would probably be a lot of work, so it's definitely more something to keep in mind for the future rather than something that makes sense to change now. It seems like the library is directly exposing Cython classes and functions to normal Python users. There's nothing wrong with this, but it means that it's pretty easy to accidentally stumble into a pretty difficult to understand/debug type errors, and also means that there's input which the library could sanitize and use that instead leads to a crash. What I personally like to do is wrap all my Cython functions in short Python wrappers than handle all the type conversions and checking.

If you wanted to see an example of this, here's a really simple Cython library I wrote a couple years back which packs integers with an arbitrary number of bits into a contiguous byte array. The Cython code is [here](https://github.com/phil-mansfield/minnow/blob/master/python/cy_bit.pyx) and here's the [wrapper Python](https://github.com/phil-mansfield/minnow/blob/master/python/bit.py).

RE: MPI

I have mpi4py 3.1.4 installed. I haven't noticed issues with it on the cluster in the past, but maybe it's related to the specific functions different programs are calling. Do you know if there's an MPI version requirement?

There's definitely something wacky doing on: it's odd for something to work on the login node but not on SLURM, but I've also never seen this happen with other MPI programs. I'll ping the cluster staff and see if they know. I'll get back to you on this if it's related to anything on your end.

RE: scipy

It makes sense that it works on an empty environment. In that case pip will just grab the latest version of scipy. The issue would be installing cosmic_profiles on a machine that already has an old version of scipy. I think you just want to figure out a version that doesn't have the problem anymore and add that as a scipy version requirement.

I foolishly didn't check what version of scipy I was running before updating it (it was about a year old), but the current version certainly works. There are some old stack overflow discussions about a similar message https://stackoverflow.com/questions/50418047/import-error-no-module-named-scipy-lib, but these predate when I installed scipy, so maybe this is a periodic problem.

RE: read-only

The issue was that I was accidentally passing a scalar for r_vir instead of a length-1 array.

I'll continue working on this can get more notes to you soon.
