### 2023-04-27 5caefa6
Removed prefixes; added convenience features for unit system 'kpc', 'kpc/h' etc; named arguments to facilitate ease of use; internal conversion from user type to float etc rather than enforcing float
