+++ cosmic_profiles/dens_profs/dens_profs_classes.pyx
@@ -6,7 +6,6 @@ import numpy as np
 from cosmic_profiles.common.cosmic_base_class cimport CosmicBase
 from cosmic_profiles.common import config
 from cosmic_profiles.common.python_routines import print_status, isValidSelection, getSubSetIdxCat
-from cosmic_profiles.dens_profs.dens_profs_tools import drawDensProfs
 from cosmic_profiles.gadget.read_fof import getFoFSHData, getPartType
 from cosmic_profiles.gadget.gen_catalogues import calcObjCat
 from cosmic_profiles.gadget import readgadget
@@ -63,7 +62,8 @@ cdef class DensProfsBase(CosmicBase):
         print_status(rank,self.start_time,'Starting getR200() with snap {0}'.format(self.SNAP))
         
         if rank == 0:
-            l_curr_over_target = 3.085678e24/config.OutUnitLength_in_cm
+            l_internal, m_internal, vel_internal = config.getLMVInternal()
+            l_curr_over_target = l_internal/config.OutUnitLength_in_cm
             return self.r200.base*l_curr_over_target
         else:
             return None
@@ -86,8 +86,9 @@ cdef class DensProfsBase(CosmicBase):
             in config.OutUnitMass_in_g
         :rtype: (N2,3) floats, (N2,) floats"""
         if rank == 0:
-            l_curr_over_target = 3.085678e24/config.OutUnitLength_in_cm
-            m_curr_over_target = 1.989e43/config.OutUnitMass_in_g
+            l_internal, m_internal, vel_internal = config.getLMVInternal()
+            l_curr_over_target = l_internal/config.OutUnitLength_in_cm
+            m_curr_over_target = m_internal/config.OutUnitMass_in_g
             return self.xyz.base*l_curr_over_target, self.masses.base*m_curr_over_target
         else:
             return None, None
@@ -108,8 +109,9 @@ cdef class DensProfsBase(CosmicBase):
             isValidSelection(obj_numbers, idx_cat_len)
             subset_idx_cat = getSubSetIdxCat(self.idx_cat.base, self.obj_size.base, obj_numbers)
             centers, ms = self._getMassesCentersBase(self.xyz.base, self.masses.base, subset_idx_cat, self.obj_size.base[obj_numbers])
-            l_curr_over_target = 3.085678e24/config.OutUnitLength_in_cm
-            m_curr_over_target = 1.989e43/config.OutUnitMass_in_g
+            l_internal, m_internal, vel_internal = config.getLMVInternal()
+            l_curr_over_target = l_internal/config.OutUnitLength_in_cm
+            m_curr_over_target = m_internal/config.OutUnitMass_in_g
             return centers*l_curr_over_target, ms*m_curr_over_target
         else:
             return None, None
@@ -249,7 +251,7 @@ cdef class DensProfs(DensProfsBase):
     ``getXYZMasses()``, ``getMassesCenters()``, ``_getMassesCenters()``, ``estDensProfs()``, 
     ``fitDensProfs()``, ``estConcentrations()``, ``plotDensProfs()``, ``getObjInfo()``."""
     
-    def __init__(self, float[:,:] xyz, float[:] masses, idx_cat, float[:] r200, str SNAP, float L_BOX, int MIN_NUMBER_PTCS, str CENTER, str VIZ_DEST, str CAT_DEST):
+    def __init__(self, xyz, masses, idx_cat, r200, SNAP, L_BOX, VIZ_DEST, CAT_DEST, MIN_NUMBER_PTCS = 200, CENTER = 'mode'):
         """
[…]
