### 2023-04-27 #5008 @tibordome
I have implemented some suggestions on the dev branch already:

  - convenience features: 'kpc', 'kpc/h' etc, plotting functionalities updated accordingly
  - named arguments for class initializers to facilitate ease of use
  - internal conversions from double to float, rather than enforcing float in the constructor
  - D_LOGSTART and D_LOGEND are now both floats

Hiding the cython functionalities is a good idea, many thanks for drawing my attention to that and your packages you mentioned. My cython classes that are exposed to the user are only nominally cython classes, the real cython functionalities are hidden. I can imagine that having a Python wrapper for each cython function might cause some overhead in terms of lines of code / readability etc, but I do now think exposing cdef classes is maybe not the smartest option. I'll try to avoid this in the future.

I'll have to check the MPI version requirement. I'll also see about scipy requirements.

I had VIZ_DEST and CAT_DEST tied to class methods in earlier versions but I found it slightly annoying to carry it around all the time, not just in the code internally (instead of self. variables of the class), but also when calling the methods as a user. Many methods would need SNAP, VIZ_DEST and CAT_DEST. I think the use case of allowing for multiple VIZ_DEST and CAT_DEST folders would be rather limited.

Speaking of use cases, I can imagine that someone would like to run two instances with different unit systems at the same time, but I think working with one simulation per time should be sufficient for most users. Not all the functions are class functions, and I am using the global unit variables fairly often (which should be avoided in general), so restructuring would take some time. Regarding different scale factors, the unit system in the code does not care about physical vs comoving units. Since it works on individual objects such as halos, the density and shape profiles vs the normalized radius remain unchanged.

In zoom-in simulations, L_BOX can be set to zero. The functions that ensure that PBCs are respected are in that case not doing anything. I am not sure I understand the offsets you mentioned.
