+++ cosmic_profiles/for_docs/common/cosmic_base_class.pyx
@@ -41,7 +41,7 @@ cdef class CosmicBase:
         :type SUFFIX: string"""
         self.SNAP = SNAP
         l_curr_over_target = config.InUnitLength_in_cm/3.085678e24
-        self.L_BOX = L_BOX # self.L_BOX must be in units of 3.085678e24 cm (Mpc/h)
+        self.L_BOX = L_BOX # self.L_BOX must be in units of 3.085678e24 cm (Mpc/h (internal length units))
         self.CENTER = CENTER
         self.VIZ_DEST = VIZ_DEST
         self.CAT_DEST = CAT_DEST
@@ -55,26 +55,26 @@ cdef class CosmicBase:
     def _getMassesCentersBase(self, float[:,:] xyz, float[:] masses, int[:] idx_cat, int[:] obj_size):
         """ Calculate total mass and centers of objects
         
-        :param xyz: positions of all simulation particles in Mpc/h
+        :param xyz: positions of all simulation particles in Mpc/h (internal length units)
         :type xyz: (N2,3) floats, N2 >> N1
-        :param masses: masses of all simulation particles in 10^10*M_sun/h
+        :param masses: masses of all simulation particles in 10^10*M_sun/h (internal mass units)
         :type masses: (N2,) floats
         :param idx_cat: contains indices of particles belonging to an object
         :type idx_cat: (N3) integers
         :param obj_size: indicates how many particles are in each object
         :type obj_size: (N1,) integers
-        :return centers, m: centers in Mpc/h and masses in 10^10*M_sun/h
+        :return centers, m: centers in Mpc/h (internal length units) and masses in 10^10*M_sun/h (internal mass units)
         :rtype: (N,3) and (N,) floats"""
         return
     
     def _getShapeCatLocalBase(self, float[:,:] xyz, float[:] masses, float[:] r200, int[:] idx_cat, int[:] obj_size, float D_LOGSTART, float D_LOGEND, int D_BINS, float IT_TOL, int IT_WALL, int IT_MIN, bint reduced, bint shell_based, str suffix):
         """ Get all relevant local shape data
         
-        :param xyz: positions of all simulation particles in Mpc/h
+        :param xyz: positions of all simulation particles in Mpc/h (internal length units)
         :type xyz: (N2,3) floats, N2 >> N1
-        :param masses: masses of all simulation particles in 10^10*M_sun/h
+        :param masses: masses of all simulation particles in 10^10*M_sun/h (internal mass units)
         :type masses: (N2,) floats
-        :param r200: each entry gives the R_200 radius of the parent halo in Mpc/h
+        :param r200: each entry gives the R_200 radius of the parent halo in Mpc/h (internal length units)
         :type r200: (N1,) floats
         :param idx_cat: contains indices of particles belonging to an object
         :type idx_cat: (N3) integers
@@ -100,8 +100,8 @@ cdef class CosmicBase:
         :type shell_based: boolean
         :param suffix: suffix for file names
         :type suffix: string
-        :return: d in Mpc/h, q, s, minor, inter, major, obj_centers in units of Mpc/h,
-            obj_masses in units of 10^10*M_sun/h
+        :return: d in Mpc/h (internal length units), q, s, minor, inter, major, obj_centers in units of Mpc/h (internal length units),
+            obj_masses in units of 10^10*M_sun/h (internal mass units)
         :rtype: 3 x (number_of_objs, D_BINS+1) float arrays, 
             3 x (number_of_objs, D_BINS+1, 3) float arrays, 
             (number_of_objs,3) float array, (number_of_objs,) float array
@@ -111,11 +111,11 @@ cdef class CosmicBase:
[…]
