### 2023-04-28 5f3349c
Fixed bug in /tests; MIN_NUMBER_DM_PTCS --> MIN_NUMBER_PTCS
