### 2023-04-28 206297f
Fixed bug in /tests; MIN_NUMBER_DM_PTCS --> MIN_NUMBER_PTCS
