### 2023-04-28 e437072
Fixed bug in /tests; MIN_NUMBER_DM_PTCS --> MIN_NUMBER_PTCS
