### 2023-04-28 #5008 @tibordome
On Linux, pip seems to be less sensitive to the scipy version that is installed, and sometimes upgrades scipy automatically,

 - Installing collected packages: scipy, cosmic-profiles
  Attempting uninstall: scipy
    Found existing installation: scipy 1.3.1
    Uninstalling scipy-1.3.1:
      Successfully uninstalled scipy-1.3.1
Successfully installed cosmic-profiles-1.3.1 scipy-1.10.1

can occur sometimes. I have added to the docs that in general scipy==1.10.1 is a requirement though pip works (no install nor runtime issues) down to scipy==1.6.1.
