### 2023-04-28 #5008 @phil-mansfield
Ah, oops, missed the comment in the documentation about L_BOX=0. That's my bad.
I agree with you about the effort required to implement some of those suggestions: I think it's fine to keep things as-is.
