### 2023-04-28 #5008 @phil-mansfield
I'll premark comments with an (S) to denote things that are pure suggestions (usually stylistic comments)

Comments related to `getShapeCatLocal`:

I encountered some problems related to translation invariance. I have a halo that I read in from my zoom-in simulations that's centred on ~[0, 0, 0] kpc with a virial radius of  ~250 kpc. When I run getShapeCatLocal on it, I get a centre of [ 0.07100291  0.01147768  0.20005868] kpc (good!), but if I add 20,000 kpc to the x, y and z positions of the halo, I get [20490.355 20480.424 20485.807] kpc instead. So the new centre I'm getting is off by about 2.5%, which ends up being well outside the virial radius of the original halo. This was with a 3 million particle halo. I tried the same trick with a 200 particle halo and its centre was translation invariant. If I set L_BOX in your example code to largeish numbers, I can get the same effect (e.g. L_BOX = 100). Can you see if you can replicate this?

If I had to guess, I'd say that the issue was related to the resolution of a 32-bit flaoting point number. The floating point mantissa has 23 bits, so you'll have a dynamic range of 1/2^24, meaning about 1-in-16-million. So positions can still be represented accurately, but sums of a large numbers of positions can't. So, for example, say you were taking the center of mass by adding up x for all 3 million particles and dividing by 3 million, after the first couple hundred thousand particles are added, every subsequent x can't be fully represented with the dynamic range of the sub. Averaged across a large population, this leads to a systematic bias. (I'm using CENTER="mode", so it probably isn't this exact issue).

I think the way you want to handle this is to make sure that you never have any variables which are raw sums of particle properties. Let me know if you'd like to talk about specific strategies that avoid this type problem. Approaches vary depending on what you want to do.

(S) I'd recommend revisiting how skipped (np < MIN_NUMBER_PTCS) haloes work. Right now, the user supplies an idx_cat array with all their haloes, but the r_vir array needs to correspond to the number of objects with np >= MIN_NUMBER_PTCS, so the user needs to manually filter on the minimum number of particles anyway. Additionally, this means that the returned array is a different length than idx_cat, meaning that if the user wants to cross-reference shape profiles with the the particles or some pre-existing halo catalogue later, the indexing becomes complicated.

I'd recommend doing one of two things instead: (1) just letting the user decide which objects they want to pass and not handling the min particle cuts yourself. (2) using rvir and idx_cat of the same length and not completely removing haloes below the particle limit but filling all their properties with NaNs. This would make sure the indexing stays easy between input and output arrays.

(S) This function returns a large number of values, many of which have similar forms to one another. I think it makes the return signature a bit overwhelming and easy to make mistakes on. (I think the user would probably need to look up the documentation every time they call it rather than being able to remember the return signature). There are a couple potential tricks you can use to clean things up a bit. The first is returning an object where the fields are the different return quantities. This is, e.g.,  what the scipy.optimize routines do (e.g. https://docs.scipy.org/doc/scipy/reference/generated/scipy.optimize.minimize_scalar.html#scipy.optimize.minimize_scalar).

Another alternative that I like is structured numpy arrays. These are numpy arrays that tie various fields together and use dictionary-like field accessing semantics. So, e.g., if you had a structured array called `shapes`, you could get the minor axis vector in the 12th shell for the 20th halo with `shapes["minor"][20,12]` and can get all the minor vectors in the 12th bin with `shapes["minor"][:,12]`, etc. One benefit of these is that you can perform filtering a single time for all the fields simultaneously, so, e.g. if you wanted to do a bunch of analysis on the 20th halo but wanted to remove unconverged bins, instead of needing to manually filter every bin, you could do something like
```
shape = shapes[20]
is_conv = np.~np.isnan(shape["d"])
shape = shape[is_conv]
```

In your case, I think you'd probably want one array for the profiles and one for the halo bulk properties. So something like this: (I haven't tested this code)
```
SHAPE_PROF_DTYPE = [("d", "f4"), ("s", "f4"), ("q", "f4"), ("minor", "f4", (3,)), ("inter", "f4", (3,), ("major", "f4", (3,)]
OBJECT_PROPERTIES_DTYPE = [("centre", "f4", (3,)), ("mass", "f4")]
shapes = np.zeros((n_pass, D_BINS+1), dtype=SHAPE_PROF_DTYPE)
objs = np.zeros((n_pass, D_BINS+1), dtype=OBJECT_PROPERTIES_DTYPE)
# etc.
```
Then the function would return
```
shapes, obj = getShapeCatLocal()
```

(S) If you do end up making this into structured arrays or an object, I'd recommend adding a field called something like "ok", which is true if the shape array is non-NaN at that index and false otherwise. It'll make people's code cleaner and a lot of people don't know how to test whether a quantity is NaN (e.g., they might write `is_converged = d != np.nan`).
