### 2023-04-29 #5008 @tibordome
Regarding translation invariance: I do not think this is due to floating point arithmetic but due to

    def respectPBCNoRef(xyz, L_BOX = None):
    """
    Return modified positions xyz_out of an object that respect the box periodicity

    If point distro xyz has particles separated in any Cartesian direction
    by more than L_BOX/2, translate those particles accordingly.

    :param xyz: coordinates of particles
    :type xyz: (N^3x3) floats
    :param L_BOX: periodicity of box (0.0 if non-periodic)
    :type L_BOX: float
    :return: updated coordinates of particles
    :rtype: (N^3x3) floats"""
    if L_BOX != 0.0:
        xyz_out = xyz.copy() # Otherwise changes would be reflected in outer scope (np.array is mutable).
        ref = 0 # Reference particle does not matter
        dist_x = xyz_out[:,0]-xyz_out[ref, 0]
        dist_y = xyz_out[:,1]-xyz_out[ref, 1]
        dist_z = xyz_out[:,2]-xyz_out[ref, 2]
        xyz_out[:,0][dist_x > L_BOX/2] = xyz_out[:,0][dist_x > L_BOX/2]-L_BOX
        xyz_out[:,0][dist_x < -L_BOX/2] = xyz_out[:,0][dist_x < -L_BOX/2]+L_BOX
        xyz_out[:,1][dist_y > L_BOX/2] = xyz_out[:,1][dist_y > L_BOX/2]-L_BOX
        xyz_out[:,1][dist_y < -L_BOX/2] = xyz_out[:,1][dist_y < -L_BOX/2]+L_BOX
        xyz_out[:,2][dist_z > L_BOX/2] = xyz_out[:,2][dist_z > L_BOX/2]-L_BOX
        xyz_out[:,2][dist_z < -L_BOX/2] = xyz_out[:,2][dist_z < -L_BOX/2]+L_BOX
        return xyz_out
    else:
        return xyz

which chooses a particle randomly (ref=0) and translates other particles so that halos at the edges of a box appear as whole objects again. If the new center reported by CosmicProfiles is outside the box, e.g. [20490.355 20480.424 20485.807], then I think I should manually ensure within the code that this is translated back into the box. Is that the suggestion?

Regarding point 2, skipping halos: Actually ``r_vir`` supplied by the user should have the same length as ``idx_cat`` (i.e. provided for non-skipped and skipped objects). All the returned quantities, on the other hand, will work with ``N_pass`` (those that have np larger or equal to ``MIN_NUMBER_PTCS``). I could provide ``len(nb_halos)`` shaped return values with lots of NaNs. However, I think determining which objects pass and which do not is a one-liner for the user. Also, the meaning of NaN would be overloaded since in addition to the non-convergence of the algorithm it could mean a not-sufficiently-resolved object. Alternatively, I could provide a functionality ``getIdxsPass()`` to provide this array upon request.

Points 3 & 4: I have adopted these suggestions on the ``dev`` branch. Very helpful indeed and it makes the return signatures much more readable. Many thanks! I was not aware of structured numpy arrays before. I think I settled on ``is_conv`` as the field that tells the user whether it converged, and ``shapes = getShapeCatLocal()`` only (rather than ``shapes, obj = getShapeCatLocal()``) , but with ``obj = getMassesCenters()`` all the necessary information can be retrieved.
