### 2023-04-29 f8455b5
Added structured numpy array for better readability of return signatures etc
