+++ example_scripts/gadget3_hdf5.py
@@ -17,13 +17,13 @@ sys.path.append(os.path.join(currentdir, '..')) # Only needed if cosmic_profiles
 from cosmic_profiles import DensShapeProfsGadget, updateInUnitSystem, updateOutUnitSystem
 
 # Parameters
-updateInUnitSystem(in_unit_length_in_cm = 3.085678e21, in_unit_mass_in_g = 1.989e43, in_unit_velocity_in_cm_per_s = 1e5)
-updateOutUnitSystem(out_unit_length_in_cm = 3.085678e24, out_unit_mass_in_g = 1.989e33, out_unit_velocity_in_cm_per_s = 1e5)
+updateInUnitSystem(length_in_cm = 'Mpc/h', mass_in_g = 'Msun/h', velocity_in_cm_per_s = 1e5, little_h = 0.6774)
+updateOutUnitSystem(length_in_cm = 'kpc/h', mass_in_g = 'Msun/h', velocity_in_cm_per_s = 1e5, little_h = 0.6774)
 CAT_DEST = "./cat"
 GROUP_DEST = "./ex_data/LLGas256b20/groups_035"
 SNAP_DEST = "./ex_data/LLGas256b20/snapdir_035/snap_035"
 VIZ_DEST = "./viz"
-D_LOGSTART = -2
+D_LOGSTART = -2.0
 D_LOGEND = 0.5
 D_BINS = 30 # If D_LOGSTART == -2 D_LOGEND == 1, 60 corresponds to shell width of 0.05 dex
 IT_TOL = np.float32(1e-2)
@@ -41,11 +41,11 @@ frac_r200 = 0.5 # At what depth to calculate e.g. histogram of triaxialities (cf
 def HDF5Ex():
     
     # Define DensShapeProfsGadget object
-    cprofiles = DensShapeProfsGadget(SNAP_DEST, GROUP_DEST, SNAP, MIN_NUMBER_PTCS, D_LOGSTART, D_LOGEND, D_BINS, IT_TOL, IT_WALL, IT_MIN, CENTER, RVIR_OR_R200, OBJ_TYPE, VIZ_DEST, CAT_DEST)
+    cprofiles = DensShapeProfsGadget(SNAP_DEST, GROUP_DEST, OBJ_TYPE, SNAP, VIZ_DEST, CAT_DEST, RVIR_OR_R200 = RVIR_OR_R200, MIN_NUMBER_PTCS = MIN_NUMBER_PTCS, D_LOGSTART = D_LOGSTART, D_LOGEND = D_LOGEND, D_BINS = D_BINS, IT_TOL = IT_TOL, IT_WALL = IT_WALL, IT_MIN = IT_MIN, CENTER = CENTER)
     
-    idx_cat = cprofiles.getIdxCat()[0] # Only rank == 0 content is of interest
+    obj_size = cprofiles.getIdxCat()[1] # Only rank == 0 content is of interest
     if rank == 0:
-        h_idx_cat_len = len(idx_cat)
+        h_idx_cat_len = len(obj_size)
     else:
         h_idx_cat_len = None
     h_idx_cat_len = comm.bcast(h_idx_cat_len, root = 0)
@@ -80,7 +80,7 @@ def HDF5Ex():
     comm.Bcast(dens_profs, root = 0)
     
     # Fit density profiles
-    best_fits = cprofiles.fitDensProfs(dens_profs[:,25:], ROverR200[25:], 'nfw', obj_numbers)
+    best_fits = cprofiles.fitDensProfs(dens_profs[:,25:], ROverR200[25:], 'nfw', obj_numbers) # Structured numpy array
 
[…]
