### 2023-05-02 #5008 @phil-mansfield
Glad to hear suggestions 3 & 4 were useful. Structured arrays are such a cool feature: they're basically very light-weight pandas dataframes: very addicting once you get used to them. (Or maybe pandas dataframes are more heavy-duty structured arrays? Not sure which came first.)

For point 1, I don't think this can be the cause of my issue. My apologies if I under-specified what I was doing. When I run this on my zoom-in halo's particles, I set L_BOX = 0.0, so this particular block of code never executed. I also don't think it can be a strict geometric issue: if I subsample my zoom-in halo to have a few hundred particles, I no longer have this issue.

I think the thing that I said which was confusing was that I can get the same problem with the example code you sent me a few days ago by changing L_BOX and the number of particles in the randomly-generated halo. The code shifts the halo into the middle of the box, so changing L_BOX is effectively just shifting the halo to larger offsets. Here are the inputs I used and what I got back:

`L_BOX = np.float32(0.0)` and `halo_res = int(3e6)`: Accurate centres
`L_BOX = np.float32(100.0)` and `halo_res = int(3e6)`: centres which are off by ~0.8, which leads to NaN shape profiles
`L_BOX = np.float32(100.0)` and `halo_res = int(3e5)`: Accurate centres
`L_BOX = np.float32(1000.0)` and `halo_res = int(3e5)`: centres which are off by about 0.3, leaning to NaN shape profiles

To me this seems consistent with floating point errors. Can you check if you can replicate this?

I think one hacky way to deal with this would be do a trick similar to what you do in `respectPBCNoRef`, select a random reference point within the halo (actually, I'd select a bunch of truly random points and take the median to make sure you get something close to the centre: it's an old trick people use when implementing quicksort), producing `delta_xyz = xyz - ref_xyz` and doing all your centre-finding and shape analysis on that. Then when you report centres back to the user, you add `ref_xyz` back.

I think that would fix the issue for any halo with randomly-ordered particles (like your synthetic halo and like the formatting used for my zoom-in haloes). I think you could still get yourself in trouble if someone had stored particles along a hilbert curve rather than randomly and had _a ton_ of particles, but I'm not sure it's worth worrying about working on such large objects unless you want to. The way you'd want to handle this is that instead of, e.g., taking averages by adding up a bunch of numbers and dividing at the end, you have a running average which is updated with each element. You could also gain some resolution by using float64 internally.

(Also, do you have some later code which makes sure that halo centers get put back within L_BOX? E.g., suppose that the L_BOX = 100, ref is at x = 99.9 and the halo centre is at x = 0.1. Does the centre get reported at x = 0.1 or as x = 100.1?)
