### 2023-05-03 1dfd38f
Updated calcCoM (delta_xyz instead of xyz); added recentreObject() to better report center values
