### 2023-05-03 c3aa9e1
promoted float32 arrays to float64 to minimise floating point arithmetic issues (occurred for several-million-particle halo and L_BOX != 0.0); averaging over ~50 random particles' position as in quicksort
