+++ cosmic_profiles/dens_profs/dens_profs_classes.pyx
@@ -22,7 +22,7 @@ cdef class DensProfsBase(CosmicBase):
     ``getXYZMasses()``, ``getMassesCenters()``, ``_getMassesCenters()``, ``estDensProfs()``, 
     ``fitDensProfs()``, ``estConcentrations()``, ``plotDensProfs()``, ``getObjInfo()``."""
     
-    def __init__(self, float[:,:] xyz, float[:] masses, int[:] idx_cat, float[:] r200, int[:] obj_size, str SNAP, float L_BOX, int MIN_NUMBER_PTCS, str CENTER, str VIZ_DEST, str CAT_DEST, str SUFFIX):
+    def __init__(self, double[:,:] xyz, double[:] masses, int[:] idx_cat, double[:] r200, int[:] obj_size, str SNAP, double L_BOX, int MIN_NUMBER_PTCS, str CENTER, str VIZ_DEST, str CAT_DEST, str SUFFIX):
         """
         :param xyz: positions of all simulation particles in Mpc/h (internal length units)
         :type xyz: (N2,3) floats, N2 >> N1
@@ -112,7 +112,7 @@ cdef class DensProfsBase(CosmicBase):
             l_internal, m_internal, vel_internal = config.getLMVInternal()
             l_curr_over_target = l_internal/config.OutUnitLength_in_cm
             m_curr_over_target = m_internal/config.OutUnitMass_in_g
-            OBJECT_PROPERTIES_DTYPE = [("centre", "f4", (3,)), ("mass", "f4")]
+            OBJECT_PROPERTIES_DTYPE = [("centre", "f8", (3,)), ("mass", "f8")]
             objs = np.zeros((len(obj_numbers),), dtype=OBJECT_PROPERTIES_DTYPE)
             objs["centre"] = centers*l_curr_over_target
             objs["mass"] = ms*m_curr_over_target
@@ -162,7 +162,7 @@ cdef class DensProfsBase(CosmicBase):
         cdef int D_LOGSTART
         cdef int D_LOGEND 
         cdef int D_BINS
-        cdef float IT_TOL
+        cdef double IT_TOL
         cdef int IT_WALL
         cdef int IT_MIN
         try:
@@ -179,7 +179,7 @@ cdef class DensProfsBase(CosmicBase):
             IT_TOL = 0.0
             IT_WALL = 0
             IT_MIN = 0
-        dens_profs = self._estDensProfsBase(self.xyz.base, self.masses.base, self.r200.base, self.idx_cat.base, self.obj_size.base, np.float32(ROverR200), obj_numbers, direct_binning, spherical, reduced, shell_based, D_LOGSTART, D_LOGEND, D_BINS, IT_TOL, IT_WALL, IT_MIN)
+        dens_profs = self._estDensProfsBase(self.xyz.base, self.masses.base, self.r200.base, self.idx_cat.base, self.obj_size.base, np.float64(ROverR200), obj_numbers, direct_binning, spherical, reduced, shell_based, D_LOGSTART, D_LOGEND, D_BINS, IT_TOL, IT_WALL, IT_MIN)
         return dens_profs
     
     def fitDensProfs(self, dens_profs, ROverR200, str method, obj_numbers): # Public Method
@@ -195,9 +195,9 @@ cdef class DensProfsBase(CosmicBase):
         :type obj_numbers: list of int
         :return: best-fits for each object
         :rtype: structured array, containing (N3, n) floats, where n is the number of free parameters in the model ``method``"""
-        best_fits = self._fitDensProfsBase(self.r200.base, self.obj_size.base, np.float32(dens_profs), np.float32(ROverR200), method, obj_numbers)
[…]
