### 2023-05-03 #5008 @tibordome
I managed to reproduce the error. I checked and removed the few apparent raw particle sums such as ``mass_total += masses[0]``. I followed your advice and improved the calculation of CoM centers by recentering with respect to some random reference points, ``delta_xyz = xyz - ref_xyz ``

    def calcCoM(xyz, masses):
        """ Calculate center of mass of point distribution

        :param xyz: coordinates of particles of type 1 or type 4
        :type xyz: (N,3) floats
        :param masses: masses of the particles
        :type masses: (N,3) floats
        :return: com, center of mass
        :rtype: (3,) floats"""
        # Average over some random particles and recentre with respect to that to avoid large numbers
        rng = default_rng(seed=0)
        choose = rng.choice(np.arange(len(xyz)), (min(50,len(xyz)),), replace = False)
        ref_xyz = np.average(xyz[choose], axis = 0)
        delta_xyz = xyz.copy()-ref_xyz
        mass_total = np.sum(masses)
        com = np.sum(np.reshape(masses, (len(masses),1))*delta_xyz/mass_total, axis = 0)
        com = com+ref_xyz
        return com

This alone could not resolve the floating point issues that accumulated with the conversion of units in the code to internal units and subsequent analysis, then back-conversion etc. What did resolve all such issues was to promote all float32 arrays to float64, float to double etc. as you suggested. I haven't looked into particles along a hilbert curve and other complicated configurations of points. I have pushed the improvements to the dev branch.

Let me know whether you have more suggestions, I can make a release then combining all the new additions.
