### 2023-05-03 1fe06c3
Vectorized raw sum in calcCoM()
