### 2023-05-05 #5008 @phil-mansfield
I am still waiting to hear back from my computing cluster IT people about the MPI issue. The first set of people I spoke to didn't know and it's been bounced around to others.

Additional thoughts:
1. (S) Personally, I would have split up the arguments to the shape profile constructors and the getShapeCat* methods differently. An alternative structure for the library would be to just have single functions that handle each of the profile tasks rather than an object+method split, but the way you've structured it allows you to split up the I/O and particle management from actually specifying what type of profile you want and how you want to calculate it. This is a good idea! It means someone could, e.g., try out different ways of calculating shapes without needing to reload the particles or resupply halo information. However, some of the arguments that specify how the profiles should be calculated are in the methods (e.g. reduced weighting) and some are in the constructor (radial bins, iteration counts). This means that if you want to tinker with the specifics of how you're calculating shape profiles (say, if I wanted to overplot a couple different binning choices to make sure I didn't have too much noise at small radii) sometimes you can just call the method again and sometimes you need to cal the constructor again (and potentially repeat all the I/O). I'd recommend putting everything that specifes the details of the profile in the method.

2. (S) Regardless of how you want to handle this, I'd recommend making the call signatures of the density profile mthods/constructors as close to the shape profiles/constructors as you can. The biggest difference I see is that radial bins are handled differently. In the density profiles, radial bins are passed in to the method and the user can specify whatever bins they'd like, while in the shape classes, the user is forced to use a particular binning scheme and has to specify it in the constructor. I'd recommend using the same technique in the shape classes.

3. (S) Is there any technical reason that the shape and density profiles can't use the same constructors? I think it would simplify things. The main thing that I see is that the density constructor doesn't need the iteration arguments, but those can be passed to the shape calculation methods as optional keyword arguments instead.

4.

> On the other hand, with direct_binning = False, we perform a kernel-based density profile estimation, cf. [Reed et al. 2005](https://academic.oup.com/mnras/article/357/1/82/1039256). Kernel-based approaches allow estimation of profiles without excessive particle noise.

Can you please say some more about what is being done with this KDE method? Reading Reed's appendix, it seems like it's a Gaussian kernel whose size is given by some power law and it's unclear to me how they decided that this was a good power law to use. KDEs are the type of thing you often need to fiddle with, so it might be good to give the user some control over the details. (Also, this is a bit pedantic on my side, but I don't think KDEs reduce the noise in density profiles. They smooth out the profile and remove the visual features that allow us to tell that the profile is noisy, but they won't change the statistical uncertainty associated with Poisson noise. This doesn't mean that they're bad to use.)

5. When citing the source papers for the different profiles you can fit, there's no year for the NFW citation and I think Zhao et al. 1996 is the right citation for the generalized NFW profile https://academic.oup.com/mnras/article/278/2/488/951933

6. (S) When fitting profiles, it might be nice to allow users to fix a parameter, e.g., people often fix alpha to some value they like when fitting Einasto profiles, or manually set one of the generalized NFW profile indices to something.

7. Can you please specify how the fits are being performed (e.g. non-linear Levenberg-Marquardt, MCMC, manual likelihood maximization, etc.) and what error model is being used (e.g. are you using uniform log-error bars, are you estimating error bars from Poisson statistics, etc.)

8. (S) For choosing an inner converged radius of a profile, I think Ludlow et al. 2019 is probably the most cutting edge reference https://ui.adsabs.harvard.edu/abs/2019MNRAS.488.3663L/abstract

9. Can you please explain the specifics of how your mock halo generation algorithm works? Based on the description, I would have guessed that it would be sampling the density profile, i.e., in the spherically symmetric case, the probability of a point being generated at r is proportional to r^2 * rho(r), (and in the ellipsoidal case, it'd proportional to r_ell^2 * rho(r_ell)) then some random uniform angles would be generated at that radius. However, the point distribution is very structured in the plots included on this page, which makes me think that something else is being done (just eyeballing it, it looks like you analytically figure out how many particles to generate at discrete r_ell's, then have some deterministic scheme for spreading them out over the ellipsoidal shell).

10. (S) The reason I'm asking about this is that personally I think most users who are interested in generating mock point distributions would want randomly generated points. One of the main things I do with mock haloes is test out the impact of Poisson noise and do sanity checks on algorithms that need to be tested on realistic point distributions. With a scheme like the one I'm guessing is being used here, neither of those things will be possible. Additionally, if particles are being put at discrete, fixed r_ell, it means that looking at the halo properties with bins that aren't exactly aligned with the generating bins will lead to weird spikes and artefacts.

(Also, I've never seen this published before but I'm pretty sure that the iterative shape routine that everyone likes to use biases to low c/a ratios for real point distributions, but not for ones generated at a fixed r_ell shells. This isn't a critique of writing a library that uses this method, since it's so wide-spread already, but it means that people trying to understand the limitations of their measurements might be misled.)

I'd recommend adding a method that generates a random halo. The way to do this would be to use inverse transfer sampling as a function of r_ell, randomly generating theta & phi within the r_ell-based coordinate system, then transforming all the points to normal cartesian coordinates. I'm guessing you're already doing the last of those three steps. You can do the first by by generating a mass uniformly at random between (0, M(r_max)), and inverting M(r_ell) to get r_ell(M) so this turns into a randomly generated r_ell that follows the correct density profile. (It's expensive to manually integrate and invert M(r_ell), so you could just do it once and interpolate). Then you can get random angles with this method: https://mathworld.wolfram.com/SpherePointPicking.html.

If you're interested in generating random point distributions, but are confused by that description, let me know. I think I have some code that does this sitting on a old laptop somewhere: I can track it down and send it to you as an example if you want.

11. (S) I think the typical physicist would be confused by what the caching code is doing. I'd recommend backing up a bit and describing it at a level that someone who hasn't written code in a memory-managed language/taken a systems class would understand.

Also, I was a bit confused by what exactly the `use_memory_up_to` argument was doing. On first read I thought it was the size of the cache, but it's actually the size of the things /not/ in the cache. Maybe I'm unfamiliar with psutil, but reading through the docs it seems like the cache size is scaled by the amount of unallocated but available memory. Is that correct? The amount of unallocated memory will change s the program runs (and as the cache expands). Does this get recalculated dynamically? Also, sometimes when running on a cluster, you can specify how much memory you want to use (which is smaller than the actual amount of RAM on the node you're running the job on) and if your program ever goes over that limit, an automated task sees that and manually kills your job. It's possible that this scheme would mean a user running code on such a setup would find it hard to limit the size of their cache correctly.

I'd recommend just having the user specify the cache size directly.
