### 2023-05-06 e5fd58e
move some arguments from constructors to method arguments; shape and density profiles use same constructors: DensShapeProfs and DensShapeProfsGadget; default_katz_config dictionary
