### 2023-05-06 8b960aa
Fixed bug in calcDensProfsKernelBased(): the kernel width normalization should be h_0.1 = 0.005r_vir at 0.1r_vir as suggested by Reed et al 2005
