### 2023-05-06 986bff6
Allow fixing of parameter in density profile fitting, by simply adding value to method dict; added corresponding checkDensFitMethod(); removed decorator around fitDensProfHelper() since dict not hashable
