### 2023-05-07 3ed6f5c
Fixed bug in test_shapes.py
