### 2023-05-07 #5008 @tibordome
1) That is a fair point. I have moved the ``ROverR200``, ``IT_TOL``, ``IT_WALL``, ``IT_MIN``, ``REDUCED``, ``SHELL_BASED`` parameters away from the constructor of the shape profiling class, bundled them into one dictionary, ``katz_config``, and transferred them into the respective methods that need them.

2) I have replaced ``D_LOGSTART``, ``D_LOGEND`` and ``D_BINS`` by ``ROverR200``, to match the density profiling class.

3) Since internally the code is relying on multiple levels of inheritance, the shape profiling class had all functionalities of the density profiling class, + shape profiling. Indeed, it was thus redundant to provide two interfaces to the user. Any profiling now uses one constructor.

4) On page 85 of Reed et al 2005, they motivate the choice of the power law: "To calculate density profiles without excessive particle noise, we
developed a novel kernel-based algorithm1 (Merritt & Tremblay
1994); see Appendix for a full description. Both the width and the
shape of the kernel are varied with radius; the variation in shape is
significant near the origin, where a symmetric kernel would ‘over-
flow’ the r = 0 boundary. The window width must be carefully
chosen to reduce Poisson noise (‘variance’) without oversmooth-
ing (‘biasing’) the profile. In general, it may be shown (e.g. Scott
1992, p. 130) that when the window width is chosen to minimize the
mean square error of the estimate, most of the error will come from
the variance. Window widths large enough to eliminate the ‘wig-
gles’ will generally bias the slope. In addition, the window width
should vary with local particle density (Abramson 1982), roughly
as ρ−1/2 . We used a kernel window width that varied as r 0.5 set at
h0.1 = 0.005r vir at 0.1r vir as it yields profiles and profile slopes in
good agreement with binned profiles created with TIPSY 2 ), and it
preserves the central cusp and major substructure."

Hence I also choose $h(r) = 0.005 \left(\frac{r}{0.1r_{\text{vir}}}\right)^{0.5}r_{\text{vir}}$. I wouldn't know over which parameter to give control to the user here.

5) Added

6) Added. E.g. if user provides ``method = {'profile': 'einasto', 'alpha': 0.18}``, then ``alpha`` will be kept fixed.

7) The fits are performed using a truncated Newton (TNC) algorithm, which was working best I thought. Error bars are simply estimated (if there is more than 1 profile) based on 25 and 75% quantiles in each shell.

8) Added.

9) I was inspired by Zemp et al 2011, https://arxiv.org/abs/1107.5582, who on page 3 mention "For the generation of a uniform distribution of points on a surface of an arbitrary-shaped ellipsoid,
which is needed for setting up an ellipsoid with a given density profile, a method as outlined in Section 2.5.5 of Rubinstein & Kroese (2007) is used. "

So in general, I create a mock halo of mass ``tot_mass`` consisting of approximately ``res`` particles. The ``model_pars`` array contains the parameters for the profile model given in ``profile``. I first determine the number of particles in each shell such that the total number of all particles is close to ``res``. Then I draw points uniformly from an ellipsoidal shell volume, centered at the origin, using ``drawUniformFromShell()``. The scheme of spreading the particles out over the shell is not deterministic.

10) I am not too sure what a real random halo would be like compared to 9). "Additionally, if particles are being put at discrete, fixed r_ell, it means that looking at the halo properties with bins that aren't exactly aligned with the generating bins will lead to weird spikes and artefacts." I do not think there would be any spikes. The mock halo is such that the constraints ``tot_mass, res, model_pars, profile, a, b, c`` provided by the user are followed as closely as possible, but there are no empty shells etc. Zemp et al 2011 has quantified (again in https://arxiv.org/abs/1107.5582) the biases incurred in the respective iterative algorithms.

I have looked at the suggested random halo generation implementation, but I cannot quite figure out what would be provided by the user (e.g. resolution of halo?, total mass of halo?, a, b, c, model_pars, profile, r_max?), and what is calculated. If I sample mass uniformly at random between (0, M(r_max)), then isn't it possible that one single particle already has all the mass of my halo of target mass M(r_max)? Also I think https://mathworld.wolfram.com/SpherePointPicking.html does not explain how to sample angles uniformly in an ellipsoidal coordinate frame. If my halo is very elongated sampling from a sphere will deposit most particles at the 'ends of the cigar' so to say.

11) I'll look into this today.
