### 2023-05-07 f5dcc12
Improved docs on caching functionalities
