### 2023-05-10 #5008 @tibordome
@phil-mansfield, many thanks for your detailed feedback

- The red / green points are merely Fibonacci-sampling ellipsoidal surfaces to highlight a certain depth in a halo

I might allow the user to provide their own r -> h mapping in some later version, I'll keep that in mind.
