### 2023-05-12 #5008 @adonath
@tibordome One minor thing: can you please replace the shortcuts like `\mnras` in the `.bib` file by the full journal names. It seems currently the journal names are not rendered correctly in the references of the paper. Thanks!
