### 2023-05-13 6f2020b
Replaced /mnras abbreviation by the full journal name
