+++ paper/paper.bib
@@ -19,7 +19,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Dome_2023,
        author = {{Dome}, Tibor and {Fialkov}, Anastasia and {Mocz}, Philip and {Sch{\"a}fer}, Bj{\"o}rn Malte and {Boylan-Kolchin}, Michael and {Vogelsberger}, Mark},
         title = "{On the cosmic web elongation in fuzzy dark matter cosmologies: Effects on density profiles, shapes, and alignments of haloes}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {dark matter, large-scale structure of Universe, cosmology: theory, Astrophysics - Cosmology and Nongalactic Astrophysics},
          year = 2023,
         month = mar,
@@ -92,7 +92,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Diemer_2022,
        author = {{Diemer}, Benedikt},
         title = "{A dynamics-based density profile for dark haloes - I. Algorithm and basic results}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {methods: numerical, dark matter, large-scale structure of Universe, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Astrophysics of Galaxies},
          year = 2022,
         month = jun,
