### 2023-05-14 26b212f
Fixed double vs float issues in gadget functionalities; fixed r_s rescaling issue in drawDensProfs(); allow user to choose between Powell, TNC and other minimization methods
