+++ cosmic_profiles/for_docs/shape_profs/shape_profs_classes.c
@@ -10,7 +10,7 @@
             "-fopenmp"
         ],
         "include_dirs": [
-            "/home/tibor/.local/lib/python3.9/site-packages/numpy/core/include",
+            "/home/tibor/anaconda3/envs/fdm_baryons/lib/python3.8/site-packages/numpy/core/include",
             "."
         ],
         "name": "shape_profs.shape_profs_classes",
