+++ cosmic_profiles/shape_profs/shape_profs_classes.pyx
@@ -450,7 +450,7 @@ cdef class DensShapeProfsBase(DensProfsBase):
 ############################################################################################################################
         
 ################################## User provides index catalogue directly ##################################################
-        
+
 ############################################################################################################################
 cdef class DensShapeProfs(DensShapeProfsBase):
     """ Class for density profile calculations

[…]
