+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 1.3.1
+current_version = 1.4.0
 
 [build_sphinx]
 source-dir = docs/source
