### 2023-05-14 #5008 @tibordome
@adonath many thanks for pointing that out. I have fixed the ``\mnras`` shortcuts, made a tagged release (version number v1.4.0) and archived the repo on Zenodo. The archive DOI would be

[![DOI](https://zenodo.org/badge/475501515.svg)](https://zenodo.org/badge/latestdoi/475501515)
