### 2023-05-15 #5008 @dfm
@phil-mansfield — I'm doing the final processing of this submission. Can you update the metadata on the Zenodo record (there should be an "Edit" button the record page if you're logged in) so that the title and author list exactly match the paper? Then it looks like we're all set!
